### 2024-06-13 #149 @editorialbot
[PRE REVIEW]: Digital Twins for Ocean Robots
**Submitting author:** @gaelforget (<a href="http://orcid.org/0000-0002-4234-056X">Gael Forget</a>)
**Repository:** https://github.com/JuliaOcean/OceanRobots.jl
**Branch with paper.md** (empty if default branch): JuliaConProceedings2023
**Version:** v0.2.0
**Managing EiC:** Mathieu Besançon

Thanks for submitting your paper to JCON @gaelforget. **Currently, there isn't a JCON editor assigned** to your paper.
