### 2024-06-17 #149 @luraess
@gaelforget Thanks for submitting your draft to the JuliaCon Proceedings! I will be the managing editor of your paper. Could you please suggest me a succinct list of relevant reviewers.

Also, your paper failed to compile in the system with following error https://github.com/JuliaCon/proceedings-papers/actions/runs/9501383592/job/26186778694 .  Please try getting it to compile - it seems to be related to missing reference or similar. To re-trigger PDF generation, use `@editorialbot generate pdf`. Thanks!
