### 2024-06-18 9ae2434
fix invalid DOIs
