+++ docs/paper/ref.bib
@@ -35,7 +35,7 @@ ISSN={2296-7745},
 
 @article{DT2024,
 	author = {National Academies of Sciences, Engineering, and Medicine},
-	doi = {https://doi.org/10.17226/26894},
+	doi = {10.17226/26894},
 	publisher = { The National Academies Press },
 	title = {{Foundational Research Gaps and Future Directions for Digital Twins}},
 	year = {2024}
@@ -136,7 +136,7 @@ journal = {Journal of Open Source Software}
 
 @article{MFF14,
 author = {McCaffrey, K and Fox-Kemper, B and Forget, G},
-doi = {http://dx.doi.org/10.1175/JPO-D-14-0023.1},
+doi = {10.1175/JPO-D-14-0023.1},
 journal = {JPO},
 pages = {1773--1793},
 title = {{Estimates of Ocean Macro-turbulence: Structure Function and Spectral Slope from Argo Profiling Floats}},
