### 2024-06-18 #149 @gaelforget
> @gaelforget Thanks for submitting your draft to the JuliaCon Proceedings! I will be the managing editor of your paper. Could you please suggest me a succinct list of relevant reviewers.

Will do. Thanks @luraess for handling this submission.

> Also, your paper failed to compile in the system with following error https://github.com/JuliaCon/proceedings-papers/actions/runs/9501383592/job/26186778694 . Please try getting it to compile - it seems to be related to missing reference or similar. To re-trigger PDF generation, use `@editorialbot generate pdf`. Thanks!

I don't understand how to reproduce the issue that needs fixing.

- The paper compiles on my laptop without problem, incl for the BibTeX part.
- The log generated by the GitHub action suggests to look at paper.log and paper.blg but I don't have access to those files.
- I tried `docker run ghcr.io/xu-cheng/texlive-full:latest` on my laptop to see if I could reproduce the environment use by the GitHub action but get an error message.

There are warnings related to BibTeX but I am not sure this is related to the error.

So how am I supposed to proceed to figure out what the error is, so I can try figure out to fix it?

Please advise
