### 2024-06-18 #149 @gaelforget
ps. I did a couple things that did not affect the outcome : 1. fixe the INVALID DOIs 2. move `docs/paper` to `paper`
