### 2024-06-18 #149 @luraess
@gaelforget seem when comparing to other recently submitted papers (e.g. https://github.com/kaipartmann/Peridynamics.jl/tree/paper_juliacon_2023/paper), you may be missing the `.latexmkrc`, `prep.rb` files in your `paper` folder https://github.com/JuliaOcean/OceanRobots.jl/tree/JuliaConProceedings2023/paper
