### 2024-06-19 #149 @gaelforget
I added the files in https://github.com/JuliaOcean/OceanRobots.jl/commit/bcd75310d7e41480cf0d12ac5c1a8c7b85d667b5

but does not seem sufficient (see https://github.com/JuliaCon/proceedings-review/issues/149#issuecomment-2177358642)
