+++ paper/paper.tex
@@ -197,10 +197,6 @@ Further development of the DTOR framework is expected to proceed in several dire
 
 A second direction to pursue is that of further integration with the Julia software stack. Ocean modeling capabilities such as \pkg{Oceananigans.jl} \cite{OceananigansJOSS}, \pkg{PlanktonIndividuals.jl} \cite{Wu2022}, \pkg{AIBECS.jl} \cite{Pasquier2022}, \pkg{OceanColorData.jl}, and \pkg{WorldOceanAtlasTools.jl} are of immediate relevance. More broadly, there is a lot of very useful work being done across of number of github organizations that DTOR could further leverage and integrate with. To list a few : \pkg{JuliaClimate}, \pkg{JuliaOcean}, \pkg{CLIMA}, \pkg{JuliaGeo}, \pkg{JuliaEarth}, \pkg{JuliaSpace}, \pkg{JuliaRobotics}, \pkg{juliadynamics}, \pkg{MakieOrg}, \pkg{PlutoOrg}, \pkg{JuliaHub}, \pkg{GenieFramework}, \pkg{JuliaStats}, \pkg{SciML}, \pkg{FluxML}, \pkg{Turing}, and \pkg{JuliaAI}. 
 
-%MITgcm : \cite{marshall1997fvi,camp-etal:04,HHG05,Forget2015a,Forget2024b}
-%OCCA / ECCO : \cite{Forget2007,Forget2010,Forget2015a,FFL15,FP15,Forget2024b}
-%More : \cite{Forget2021,Wu2022,OceananigansJOSS,Makie,Innes2018,Pluto}
-
 \section{Science Applications}
 
 Our current focus is on geospatial analyses that track global warming and marine heatwaves using Argo data \cite{Forget2024a}. Another example from our research is with the CStream research program (UK-US) which is releasing drifters near Florida Strait to track the nutrient stream that is associated with the Gulf Stream (Fig.~\ref{fig:Cstream}). Other applications at the scale of oceanic basins include tracking ocean plastic pollution, monitoring biological impacts of extreme events and global warming, or the optimization of the global climate monitoring fleet of ocean robots for e.g. cost. More locally, DTOR could be used in the context of field experiments such as S-MODE, to choose an example from the recent past. Related projects on digital twins, not focused on exploiting Julia, include the EU funded {\it Destination Earth} program, Mercator Ocean International's {\it Digital Twin Ocean}, and the UN Decade Action's DITTO initiative ({\it Digital Twins of The Ocean}).
@@ -215,6 +211,6 @@ Our current focus is on geospatial analyses that track global warming and marine
 
 Support from NASA awards 80NSSC20K0796, 80NSSC23K0355, 80NSSC22K1697, 1676067, and 1686358, as well as from the Simons Foundation via the CBIOMES and SCOPE-GRADIENTS program is acknowledged for this work.
 
-\input{bib.tex}
+%\input{bib.tex}
 
 \end{document}
