### 2024-06-19 #149 @gaelforget
[Peridynamics.jl](https://github.com/kaipartmann/Peridynamics.jl)
