### 2024-06-19 6378ee3
try half of .bib
