+++ paper/paper.tex
@@ -211,6 +211,6 @@ Our current focus is on geospatial analyses that track global warming and marine
 
 Support from NASA awards 80NSSC20K0796, 80NSSC23K0355, 80NSSC22K1697, 1676067, and 1686358, as well as from the Simons Foundation via the CBIOMES and SCOPE-GRADIENTS program is acknowledged for this work.
 
-%\input{bib.tex}
+\input{bib.tex}
 
 \end{document}
