+++ paper/ref.bib
@@ -110,146 +110,3 @@ DOI = {10.5194/gmd-9-1937-2016}
 }
 
 
-@ARTICLE{Roemmich2019,
- AUTHOR={Roemmich, Dean and Alford, Matthew H. and Claustre, Hervé and Johnson, Kenneth and King, Brian and Moum, James and Oke, Peter and Owens, W. Brechner and Pouliquen, Sylvie and Purkey, Sarah and Scanderbeg, Megan and Suga, Toshio and Wijffels, Susan and Zilberman, Nathalie and Bakker, Dorothee and Baringer, Molly and Belbeoch, Mathieu and Bittig, Henry C. and Boss, Emmanuel and Calil, Paulo and Carse, Fiona and Carval, Thierry and Chai, Fei and Conchubhair, Diarmuid and d’Ortenzio, Fabrizio and Dall’Olmo, Giorgio and Desbruyeres, Damien and Fennel, Katja and Fer, Ilker and Ferrari, Raffaele and Forget, Gael and Freeland, Howard and Fujiki, Tetsuichi and Gehlen, Marion and Greenan, Blair and Hallberg, Robert and Hibiya, Toshiyuki and Hosoda, Shigeki and Jayne, Steven and Jochum, Markus and Johnson, Gregory C. and Kang, KiRyong and Kolodziejczyk, Nicolas and Körtzinger, Arne and Traon, Pierre-Yves Le and Lenn, Yueng-Djern and Maze, Guillaume and Mork, Kjell Arne and Morris, Tamaryn and Nagai, Takeyoshi and Nash, Jonathan and Garabato, Alberto Naveira and Olsen, Are and Pattabhi, Rama Rao and Prakash, Satya and Riser, Stephen and Schmechtig, Catherine and Schmid, Claudia and Shroyer, Emily and Sterl, Andreas and Sutton, Philip and Talley, Lynne and Tanhua, Toste and Thierry, Virginie and Thomalla, Sandy and Toole, John and Troisi, Ariel and Trull, Thomas W. and Turton, Jon and Velez-Belchi, Pedro Joaquin and Walczowski, Waldemar and Wang, Haili and Wanninkhof, Rik and Waterhouse, Amy F. and Waterman, Stephanie and Watson, Andrew and Wilson, Cara and Wong, Annie P. S. and Xu, Jianping and Yasuda, Ichiro},   
- TITLE={On the Future of Argo: A Global, Full-Depth, Multi-Disciplinary Array},      
-JOURNAL={Frontiers in Marine Science},      
-VOLUME={6},      
-YEAR={2019},      
-URL={https://www.frontiersin.org/article/10.3389/fmars.2019.00439}, 
-DOI={10.3389/fmars.2019.00439}, 
-ISSN={2296-7745},   
-}
-
-@article{Pasquier2022,
-doi = {10.21105/joss.03814},
-url = {https://doi.org/10.21105/joss.03814},
-year = {2022}, 
-publisher = {The Open Journal}, 
-volume = {7}, 
-number = {69}, 
-pages = {3814}, 
-author = {Benoît Pasquier and François W. Primeau and Seth G. John}, 
-title = {AIBECS.jl: A tool for exploring global marine biogeochemical cycles.}, 
-journal = {Journal of Open Source Software}
-}
-
-@article{MFF14,
-author = {McCaffrey, K and Fox-Kemper, B and Forget, G},
-doi = {10.1175/JPO-D-14-0023.1},
-journal = {JPO},
-pages = {1773--1793},
-title = {{Estimates of Ocean Macro-turbulence: Structure Function and Spectral Slope from Argo Profiling Floats}},
-volume = {45},
-year = {2015}
-}
-
-
[…]
