+++ paper/ref.bib
@@ -1,47 +1,3 @@
 
-@ARTICLE{Bittig2018,
-AUTHOR={Bittig, Henry C. and Steinhoff, Tobias and Claustre, Hervé and Fiedler, Björn and Williams, Nancy L. and Sauzède, Raphaëlle and Körtzinger, Arne and Gattuso, Jean-Pierre},   
-TITLE={An Alternative to Static Climatologies: Robust Estimation of Open Ocean CO2 Variables and Nutrient Concentrations From T, S, and O2 Data Using Bayesian Neural Networks},      
-JOURNAL={Frontiers in Marine Science},     
-VOLUME={5},          
-YEAR={2018},      
-URL={https://www.frontiersin.org/articles/10.3389/fmars.2018.00328},       
-DOI={10.3389/fmars.2018.00328},      
-	ISSN={2296-7745},   
-}
-
-@ARTICLE{Sauzède2017,
-  AUTHOR={Sauzède, Raphaëlle and Bittig, Henry C. and Claustre, Hervé and Pasqueron de Fommervault, Orens and Gattuso, Jean-Pierre and Legendre, Louis and Johnson, Kenneth S.},    
-TITLE={Estimates of Water-Column Nutrient Concentrations and Carbonate System Parameters in the Global Ocean: A Novel Approach Based on Neural Networks},      
-JOURNAL={Frontiers in Marine Science},      
-VOLUME={4},           
-YEAR={2017},      
-URL={https://www.frontiersin.org/articles/10.3389/fmars.2017.00128},      
-DOI={10.3389/fmars.2017.00128},      
-ISSN={2296-7745},   
-   }
-   
-@article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  doi={10.1137/141000671},
-  journal={SIAM review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM}
-}
-
-@article{DT2024,
-	author = {National Academies of Sciences, Engineering, and Medicine},
-	doi = {10.17226/26894},
-	publisher = { The National Academies Press },
-	title = {{Foundational Research Gaps and Future Directions for Digital Twins}},
-	year = {2024}
-}
-
-@article{Makie, doi = {10.21105/joss.03349}, url = {https://doi.org/10.21105/joss.03349}, year = {2021}, publisher = {The Open Journal}, volume = {6}, number = {65}, pages = {3349}, author = {Simon Danisch and Julius Krumbiegel}, title = {Makie.jl: Flexible high-performance data visualization for Julia}, journal = {Journal of Open Source Software} }
-
 @article{Innes2018, doi = {10.21105/joss.00602}, url = {https://doi.org/10.21105/joss.00602}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {25}, pages = {602}, author = {Mike Innes}, title = {Flux: Elegant machine learning with Julia}, journal = {Journal of Open Source Software} }
 
