### 2024-06-19 56597a3
rm accents in ref.bib
