### 2024-06-19 11517cf
rm accents in ref.bib
