+++ paper/paper.tex
@@ -167,7 +167,7 @@ Let's define digital twins as agreed upon by the U.S. Committee on Foundational 
 A digital twin is a set of virtual information constructs that mimics the structure, context, and behavior of a natural, engineered, or social system (or system-of-systems), is dynamically updated with data from its physical twin, has a predictive capability, and informs decisions that realize value. The bidirectional interaction between the virtual and the physical is central to the digital twin.
 \end{defi}
 
-In the {\it digital twins for ocean robots} framework (DTOR), interactivity is facilitated by Julia and its large ecosystem of software packages. In particular, \pkg{Makie.jl} \cite{Makie} and \pkg{Pluto.jl} \cite{Pluto} let us operate even complex modeling workflows from notebooks and apps. The tutorial examples in \pkg{OceanRobots.jl}, \pkg{ArgoData.jl}, \pkg{ClimateModels.jl}, and \pkg{MITgcm.jl} \cite{Forget2024b} are notably available as Pluto notebooks. Interaction between model and data is further enabled by the rich Julia ecosystem for machine learning, data assimilation, parameter inference, and so on. Fig.~\ref{fig:MLP} provides an example where we used \pkg{Flux.jl} \cite{Innes2018} to train a neural network to predict chlorophyll concentration (linked to marine microbe abundance) from environmental parameters as done in the CANYON model \cite{Sauzède2017,Bittig2018}.
+In the {\it digital twins for ocean robots} framework (DTOR), interactivity is facilitated by Julia and its large ecosystem of software packages. In particular, \pkg{Makie.jl} \cite{Makie} and \pkg{Pluto.jl} \cite{Pluto} let us operate even complex modeling workflows from notebooks and apps. The tutorial examples in \pkg{OceanRobots.jl}, \pkg{ArgoData.jl}, \pkg{ClimateModels.jl}, and \pkg{MITgcm.jl} \cite{Forget2024b} are notably available as Pluto notebooks. Interaction between model and data is further enabled by the rich Julia ecosystem for machine learning, data assimilation, parameter inference, and so on. Fig.~\ref{fig:MLP} provides an example where we used \pkg{Flux.jl} \cite{Innes2018} to train a neural network to predict chlorophyll concentration (linked to marine microbe abundance) from environmental parameters as done in the CANYON model \cite{Sauzede2017,Bittig2018}.
 
 The core of DTOR is formed by \pkg{OceanRobots.jl}, \pkg{ArgoData.jl}, and \pkg{MeshArrays.jl} for the physical twins, along with \pkg{ClimateModels.jl}, \pkg{MITgcm.jl}, and \pkg{IndividualDisplacements.jl} for the virtual twins (i.e., predictive modeling). The \pkg{ClimateModels.jl} interface streamlines the use of models implemented in various languages. Our current model hierarchy supported via \pkg{ClimateModels.jl} includes, but is not limited to, the models used in this paper. DTOR can notably simulate observations in the future based on climate model predictions using \pkg{ClimateModels.jl}, which provides two options for this kind of applications -- either querying the CMIP6 archive of model output \cite{CMIP6} or using a fast emulator such as Hector \cite{Hector2015} (Fig.~\ref{fig:globalwarming}). 
 
