### 2024-06-19 #149 @luraess
It now partly seems to build, with missing refs I guess. Did you manage to get any further insights @gaelforget ?
