### 2024-06-19 #149 @gaelforget
> It now partly seems to build, with missing refs I guess. Did you manage to get any further insights @gaelforget ?

Yes, build completed without the refs or with one ref. So I'll need to keep probing to find which ones may be the issue. Adding back a few at a time & call editorialbot each time I guess.
