### 2024-06-19 #149 @luraess
I guess this may be the most efficient for now. Unfortunately, I also have an ARM machine which would hit similar issues with running the Docker image.
