+++ paper/ref.bib
@@ -1,6 +1,6 @@
 
 @ARTICLE{Bittig2018,
-AUTHOR={Bittig, Henry C. and Steinhoff, Tobias and Claustre, Herve and Fiedler, Bjorn and Williams, Nancy L. and Sauzede, Raphaëlle and Kortzinger, Arne and Gattuso, Jean-Pierre},   
+AUTHOR={Bittig, Henry C. and Steinhoff, Tobias and Claustre, Hervé and Fiedler, Björn and Williams, Nancy L. and Sauzède, Raphaëlle and Körtzinger, Arne and Gattuso, Jean-Pierre},   
 TITLE={An Alternative to Static Climatologies: Robust Estimation of Open Ocean CO2 Variables and Nutrient Concentrations From T, S, and O2 Data Using Bayesian Neural Networks},      
 JOURNAL={Frontiers in Marine Science},     
 VOLUME={5},          
@@ -10,8 +10,8 @@ DOI={10.3389/fmars.2018.00328},
 	ISSN={2296-7745},   
 }
 
-@ARTICLE{Sauzede2017,
-  AUTHOR={Sauzede, Raphaëlle and Bittig, Henry C. and Claustre, Herve and Pasqueron de Fommervault, Orens and Gattuso, Jean-Pierre and Legendre, Louis and Johnson, Kenneth S.},    
+@ARTICLE{Sauzède2017,
+  AUTHOR={Sauzède, Raphaëlle and Bittig, Henry C. and Claustre, Hervé and Pasqueron de Fommervault, Orens and Gattuso, Jean-Pierre and Legendre, Louis and Johnson, Kenneth S.},    
 TITLE={Estimates of Water-Column Nutrient Concentrations and Carbonate System Parameters in the Global Ocean: A Novel Approach Based on Neural Networks},      
 JOURNAL={Frontiers in Marine Science},      
 VOLUME={4},           
@@ -44,212 +44,3 @@ ISSN={2296-7745},
 @article{Makie, doi = {10.21105/joss.03349}, url = {https://doi.org/10.21105/joss.03349}, year = {2021}, publisher = {The Open Journal}, volume = {6}, number = {65}, pages = {3349}, author = {Simon Danisch and Julius Krumbiegel}, title = {Makie.jl: Flexible high-performance data visualization for Julia}, journal = {Journal of Open Source Software} }
 
 @article{Innes2018, doi = {10.21105/joss.00602}, url = {https://doi.org/10.21105/joss.00602}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {25}, pages = {602}, author = {Mike Innes}, title = {Flux: Elegant machine learning with Julia}, journal = {Journal of Open Source Software} }
-
-@software{Pluto,
-  author       = {Fons van der Plas and collaborators},
-  title        = {fonsp/Pluto.jl: v0.19.42},
-  month        = may,
-  year         = 2024,
-  publisher    = {Zenodo},
-  version      = {v0.19.42},
-  doi          = {10.5281/zenodo.11144554},
-  url          = {https://doi.org/10.5281/zenodo.11144554}
-}
-
-@article{OceananigansJOSS,
-  doi = {10.21105/joss.02018},
-  url = {https://doi.org/10.21105/joss.02018},
-  year = {2020},
-  publisher = {The Open Journal},
-  volume = {5},
-  number = {53},
-  pages = {2018},
-  author = {Ali Ramadhan and Gregory LeClaire Wagner and Chris Hill and Jean-Michel Campin and Valentin Churavy and Tim Besard and Andre Souza and Alan Edelman and Raffaele Ferrari and John Marshall},
-  title = {{Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs}},
-  journal = {Journal of Open Source Software}
-}
-
[…]
