+++ paper/ref.bib
@@ -32,3 +32,42 @@ ISSN={2296-7745},
   year={2017},
   publisher={SIAM}
 }
+
+@article{DT2024,
+	author = {National Academies of Sciences, Engineering, and Medicine},
+	doi = {10.17226/26894},
+	publisher = { The National Academies Press },
+	title = {{Foundational Research Gaps and Future Directions for Digital Twins}},
+	year = {2024}
+}
+
+@article{Makie, doi = {10.21105/joss.03349}, url = {https://doi.org/10.21105/joss.03349}, year = {2021}, publisher = {The Open Journal}, volume = {6}, number = {65}, pages = {3349}, author = {Simon Danisch and Julius Krumbiegel}, title = {Makie.jl: Flexible high-performance data visualization for Julia}, journal = {Journal of Open Source Software} }
+
+@article{Innes2018, doi = {10.21105/joss.00602}, url = {https://doi.org/10.21105/joss.00602}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {25}, pages = {602}, author = {Mike Innes}, title = {Flux: Elegant machine learning with Julia}, journal = {Journal of Open Source Software} }
+
+@software{Pluto,
+  author       = {Fons van der Plas and collaborators},
+  title        = {fonsp/Pluto.jl: v0.19.42},
+  month        = may,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {v0.19.42},
+  doi          = {10.5281/zenodo.11144554},
+  url          = {https://doi.org/10.5281/zenodo.11144554}
+}
+
+@article{OceananigansJOSS,
+  doi = {10.21105/joss.02018},
+  url = {https://doi.org/10.21105/joss.02018},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {53},
+  pages = {2018},
+  author = {Ali Ramadhan and Gregory LeClaire Wagner and Chris Hill and Jean-Michel Campin and Valentin Churavy and Tim Besard and Andre Souza and Alan Edelman and Raffaele Ferrari and John Marshall},
+  title = {{Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs}},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Wu2022, doi = {10.21105/joss.04207}, url = {https://doi.org/10.21105/joss.04207}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {73}, pages = {4207}, author = {Zhen Wu and Ga{\"{e}}l Forget}, title = {PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model}, journal = {Journal of Open Source Software} }
+
