+++ paper/ref.bib
@@ -36,10 +36,8 @@ ISSN={2296-7745},
 
 
 @article{DT2024,
-	author = {National Academies of Sciences, Engineering, and Medicine},
-	doi = {10.17226/26894},
 	publisher = { The National Academies Press },
-	title = {Foundational Research Gaps and Future Directions for Digital Twins},
+	title = {{Foundational Research Gaps and Future Directions for Digital Twins}},
 	year = {2024}
 }
 
