+++ paper/ref.bib
@@ -33,15 +33,11 @@ ISSN={2296-7745},
   publisher={SIAM}
 }
 
-@BOOK{DT2024,
-  author    = "National Academy of Engineering and National Academies of Sciences, Engineering, and Medicine",
-  title     = "Foundational Research Gaps and Future Directions for Digital Twins",
-  isbn      = "978-0-309-70042-9",
-  doi       = "10.17226/26894",
-  abstract  = "Across multiple domains of science, engineering, and medicine, excitement is growing about the potential of digital twins to transform scientific research, industrial practices, and many aspects of daily life. A digital twin couples computational models with a physical counterpart to create a system that is dynamically updated through bidirectional data flows as conditions change. Going beyond traditional simulation and modeling, digital twins could enable improved medical decision-making at the individual patient level, predictions of future weather and climate conditions over longer timescales, and safer, more efficient engineering processes. However, many challenges remain before these applications can be realized.\nThis report identifies the foundational research and resources needed to support the development of digital twin technologies.  The report presents critical future research priorities and an interdisciplinary research agenda for the field, including how federal agencies and researchers across domains can best collaborate.  ",
-  url       = "https://nap.nationalacademies.org/catalog/26894/foundational-research-gaps-and-future-directions-for-digital-twins",
-  year      = 2024,
-  publisher = "The National Academies Press",
-  address   = "Washington, DC"
+@article{DT2024,
+	title = {{Foundational Research Gaps and Future Directions for Digital Twins}},
+	author = {National Academies of Sciences Engineering and Medicine},
+	doi = {10.17226/26894},
+	publisher = { The National Academies Press },
+	journal = { The National Academies Press },
+	year = {2024}
 }
-
