+++ paper/ref.bib
@@ -34,10 +34,13 @@ ISSN={2296-7745},
 }
 
 @article{DT2024,
-	title = {{Foundational Research Gaps and Future Directions for Digital Twins}},
-	author = {National Academies of Sciences Engineering and Medicine},
-	doi = {10.17226/26894},
-	publisher = { The National Academies Press },
-	journal = { The National Academies Press },
-	year = {2024}
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  doi={10.1137/141000671},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM}
 }
