+++ paper/ref.bib
@@ -52,6 +52,7 @@ ISSN={2296-7745},
   month        = may,
   year         = 2024,
   publisher    = {Zenodo},
+  journal    = {Zenodo},
   version      = {v0.19.42},
   doi          = {10.5281/zenodo.11144554},
   url          = {https://doi.org/10.5281/zenodo.11144554}
