+++ paper/ref.bib
@@ -34,7 +34,7 @@ ISSN={2296-7745},
 }
 
 @article{DT2024,
-	author = {National Academies of Sciences, Engineering, and Medicine},
+	author = {National Academies of Sciences Engineering and Medicine},
 	doi = {10.17226/26894},
 	publisher = { The National Academies Press },
 	journal = { The National Academies Press },
