+++ paper/ref.bib
@@ -41,3 +41,19 @@ ISSN={2296-7745},
 	title = {{Foundational Research Gaps and Future Directions for Digital Twins}},
 	year = {2024}
 }
+
+@article{Makie, doi = {10.21105/joss.03349}, url = {https://doi.org/10.21105/joss.03349}, year = {2021}, publisher = {The Open Journal}, volume = {6}, number = {65}, pages = {3349}, author = {Simon Danisch and Julius Krumbiegel}, title = {Makie.jl: Flexible high-performance data visualization for Julia}, journal = {Journal of Open Source Software} }
+
+@article{Innes2018, doi = {10.21105/joss.00602}, url = {https://doi.org/10.21105/joss.00602}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {25}, pages = {602}, author = {Mike Innes}, title = {Flux: Elegant machine learning with Julia}, journal = {Journal of Open Source Software} }
+
+@software{Pluto,
+  author       = {Fons van der Plas and collaborators},
+  title        = {fonsp/Pluto.jl: v0.19.42},
+  month        = may,
+  year         = 2024,
+  publisher    = {Zenodo},
+  journal    = {Zenodo},
+  version      = {v0.19.42},
+  doi          = {10.5281/zenodo.11144554},
+  url          = {https://doi.org/10.5281/zenodo.11144554}
+}
