+++ paper/ref.bib
@@ -52,8 +52,205 @@ ISSN={2296-7745},
   month        = may,
   year         = 2024,
   publisher    = {Zenodo},
-  journal    = {Zenodo},
   version      = {v0.19.42},
   doi          = {10.5281/zenodo.11144554},
   url          = {https://doi.org/10.5281/zenodo.11144554}
 }
+
+@article{OceananigansJOSS,
+  doi = {10.21105/joss.02018},
+  url = {https://doi.org/10.21105/joss.02018},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {53},
+  pages = {2018},
+  author = {Ali Ramadhan and Gregory LeClaire Wagner and Chris Hill and Jean-Michel Campin and Valentin Churavy and Tim Besard and Andre Souza and Alan Edelman and Raffaele Ferrari and John Marshall},
+  title = {{Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs}},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Wu2022, doi = {10.21105/joss.04207}, url = {https://doi.org/10.21105/joss.04207}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {73}, pages = {4207}, author = {Zhen Wu and Ga{\"{e}}l Forget}, title = {PlanktonIndividuals.jl: A GPU supported individual-based phytoplankton life cycle model}, journal = {Journal of Open Source Software} }
+
+@article{Forget2021,
+  doi = {10.21105/joss.02813},
+  url = {https://doi.org/10.21105/joss.02813},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {60},
+  pages = {2813},
+  author = {Ga{\"{e}}l Forget},
+  title = {IndividualDisplacements.jl: a Julia package to simulate and study particle displacements within the climate system},
+  journal = {Journal of Open Source Software}
+}
+
+@Article{Hector2015,
+AUTHOR = {Hartin, C. A. and Patel, P. and Schwarber, A. and Link, R. P. and Bond-Lamberty, B. P.},
+TITLE = {A simple object-oriented and open-source model for scientific and policy analyses of the global climate system – Hector v1.0},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {8},
+YEAR = {2015},
+NUMBER = {4},
+PAGES = {939--955},
+URL = {https://gmd.copernicus.org/articles/8/939/2015/},
+DOI = {10.5194/gmd-8-939-2015}
+}
+
+@Article{CMIP6,
+AUTHOR = {Eyring, V. and Bony, S. and Meehl, G. A. and Senior, C. A. and Stevens, B. and Stouffer, R. J. and Taylor, K. E.},
+TITLE = {Overview of the Coupled Model Intercomparison Project Phase 6 (CMIP6)
+experimental design and organization},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {9},
+YEAR = {2016},
+NUMBER = {5},
+PAGES = {1937--1958},
+URL = {https://gmd.copernicus.org/articles/9/1937/2016/},
+DOI = {10.5194/gmd-9-1937-2016}
+}
+
+
+@ARTICLE{Roemmich2019,
[…]
