### 2024-06-19 #149 @gaelforget
> I guess this may be the most efficient for now. Unfortunately, I also have an ARM machine which would hit similar issues with running the Docker image.

Done. Commas in the name of one ref apparently were confusing some parser somewhere.

ps. my friend had a huge laugh when learning that I just spent an hour fixing a tex build somewhere in the cloud for no good reason. Not sure who imposed tex for JCON but imho the JOSS process did not need messing up like this. Next time I would much prefer just sending you the pdf directly, compiled with tex locally on my end. Or please give user directions to use a docker instance that works on laptops, like JOSS does.
