### 2024-07-02 #149 @gaelforget
> At this stage, it would be helpful if you could suggest some reviewers in order to start the revision process.

How many do you need? Do they have to be on GitHub? Are they expected to review Julia code or just the paper itself?
