### 2024-07-02 #149 @gaelforget
Here is a list of potential reviewers that came to mind. Loosely sorted by categories.

on the Julia side :

- Lazaro Alonso (MPI)
- Martijn Visser (Deltares)
- Alexander Barth (Liege Uni)
- Chris Rackauckas (MIT)
- Greg Wagner (MIT)

on the ocean robots and digital twins side :

- Guillaume Maze (Ifremer)
- Mathieu BELBEOCH (Oceanops)
- Andrew Thompson (Caltech)
- Alison Gray (UW)
- Shane Elipot (Uni Miami)
- Louise Rousselet (MIO)
- Antoine Poteau (LOV)
- Raphaëlle Sauzède (LOV)
- Robert Todd (WHOI)
- Verena Horman (UCSD)
- João Tasso de Figueiredo Borges de Sousa (Uni. Porto)

on the digital twins side :

- Erik van Sebille (Utrecht university)
- Thomas Huang (NASA)
- Tom Farrar (WHOI)
- Mohammad Ashkezari (UW)
- Denis Felikson (NASA)
- Maike sonnewald (UC Davis)
- Martin Visbeck (Kiel Uni)
- Douglas Rao (NOAA)
- Kanna Rajan (Norwegian University of Science and Technology)
- Pierre L'hermusiaux (MIT)
