### 2024-07-05 #149 @gaelforget
> > How many do you need? Do they have to be on GitHub? Are they expected to review Julia code or just the paper itself?
>
> Thank you for providing a list. Ideally, they are familiar with Julia and the topic as we would like them to asses both the relevance of the work, but with emphasis on the related Julia code. If you see 2-3 reviewers that would fulfil these, please flag them to me further. Else, I will try to reach out to some of them in the coming days.

Thanks! Here is 5 that are are familiar with both Julia and the topic. I added their github handle (without the @).

- Jonathan Lilly (PSI, jonathanlilly)
- Alexander Barth (Liege Uni, Alexander-Barth)
- Iga Szczesniak (AIR Centre, igaszczesniak)
- Ali Ramadhan (atpdeth MRV, ali-ramadhan)
- Greg Wagner (MIT, glwagner)
