### 2024-07-16 #164 @editorialbot
[REVIEW]: Digital Twins for Ocean Robots
**Submitting author:** @gaelforget (<a href="http://orcid.org/0000-0002-4234-056X">Gael Forget</a>)
**Repository:** https://github.com/JuliaOcean/OceanRobots.jl
**Branch with paper.md** (empty if default branch): JuliaConProceedings2023
**Version:** v0.2.9
