### 2024-07-26 #164 @Alexander-Barth
Hi  @gaelforget, this is a nice paper!

I try to run the first code snipped on the your paper (Code 1).

```julia
using OceanRobots , ArgoData , CairoMakie
files_list = GDAC.files_list() ; wmo = 6900900
argo = read(ArgoFloat() , wmo = wmo , files_list = lst )
fig = plot(argo , option = : standard , pol = pol )
```

It it seems that `lst` should be `files_list`. Also `pol` is undefined.
Should such issues be discussed here, or rather directly in https://github.com/JuliaOcean/OceanRobots.jl ?
