### 2024-07-26 #164 @Alexander-Barth
Concerning "Does installation proceed as outlined in the documentation?": I did not see any specific instruction (or did I miss them?), but I can confirm that package can be installed as any julia package.
