### 2024-07-26 #164 @Alexander-Barth
I am trying to run the notebook Buoy_NWP_NOAA.jl, but I cannot change the station number. It will always show 41046 even if I select 44065.

I am wondering if there is not an issue with this line:

https://github.com/JuliaOcean/OceanRobots.jl/blob/c5179c048d504366ba2175161759261a4242ceb7/examples/Buoy_NWP_NOAA.jl#L69

Should it not be: `buoy=read(NOAAbuoy(),sta)` (rather than hard coding the station number) ?
