### 2024-07-26 #164 @Alexander-Barth
The notebook [Buoy_NWP_NOAA_monthly.jl](https://juliaocean.github.io/OceanRobots.jl/stable/examples/Buoy_NWP_NOAA_monthly.html): can you add unit to the x and y axis? It it in °Fahrenheit ?
