### 2024-07-28 #164 @luraess
> Should such issues be discussed here, or rather directly in https://github.com/JuliaOcean/OceanRobots.jl ?

@Alexander-Barth Issues specific to the code, OceanRobots.jl, could be directly discussed in the respective repo (as issue) which can reference this Review issue. Issues, comments and suggestions related to the manuscript and the revision should be discussed here. Hope this helps - thanks!
