### 2024-08-02 e35854e
NOAA : get buoy lists, improve notebook
