+++ test/runtests.jl
@@ -59,8 +59,12 @@ using Test
 
     ##
 
-    stations=[41046, 44065]		
-	NOAA.download(stations)
+    allstations=NOAA.station_list()
+    metstations=NOAA.list_realtime(ext=:txt)
+
+    stations=metstations[1:200:end]
+    ids=NOAA.download(stations)
+
     b=read(NOAAbuoy(),41046)
     plot(b,"PRES")
     @test isa(b,NOAAbuoy)
