### 2024-08-02 #164 @gaelforget
updated code snippet in https://github.com/JuliaCon/proceedings-review/issues/164#issuecomment-2265768841

addresses https://github.com/JuliaCon/proceedings-review/issues/164#issuecomment-2252988030

Good points! Thank you
