+++ dev/examples/Buoy_NWP_NOAA.jl
@@ -1,5 +1,5 @@
 ### A Pluto.jl notebook ###
-# v0.19.42
+# v0.19.45
 
 using Markdown
 using InteractiveUtils
@@ -21,37 +21,33 @@ begin
 	"Done with packages"
 end
 
-# ╔═╡ 6b5ffee8-a230-42d4-8c49-ff14e65f52b7
-	TableOfContents()
-
 # ╔═╡ d6feeecc-87db-4bfd-8341-adede2df8ea0
 begin
 	img_url="https://nosc.noaa.gov/OSC/images/Kukui_and_NOAA_buoy_1.177308.JPG"
 	md"""# NOAA Buoys
-	
-	$(Resource(img_url,:width => 250))
 
-	!!! note 
-	    For more information about NOAA Buoys, visit <https://www.ndbc.noaa.gov/>
+	Visualize real time data for buoys that report `Standard Meteorological Data`.
 	
+	!!! note 
+	    For more information about NOAA Buoys, visit <https://www.ndbc.noaa.gov/> and links below.
+
+	$(Resource(img_url,:width => 250))
 	"""
 end
 
-# ╔═╡ 8b610dbe-f9ab-4498-8c98-fc8e83474981
-begin
-	md"""## Visualize Data"""
-end
+# ╔═╡ ee3373bb-a848-4919-9163-384b6f33632a
+TableOfContents()
 
 # ╔═╡ 46ff94e2-1b1e-454d-ab92-89be94123a47
-md"""## Appendix
-
-### Code and Data
+md"""## Code and Data
 """
 
 # ╔═╡ db662fb4-7413-11ec-1af6-43b18c0c15a9
 begin
 	#See https://www.ndbc.noaa.gov/
-    stations=[41046, 44065] #, 44097, 44013]		
+    stations=[41046, 44065,  
+		31005, 41047, 41117, 42058, 44027, 44084, 45002, 45136, 45164,
+		45206, 46028, 46077, 46146, 46237, 46277, 51208, 62121, 64045]	
 	NOAA.download(stations)
    	"Done with downloading data"
 end
@@ -60,27 +56,25 @@ end
 begin
 	sta_b = @bind sta Select(stations, default=stations[1])
 	myurl="https://www.ndbc.noaa.gov/station_page.php?station=$(sta)"
-	md"""Select buoy : $(sta_b)
-    """
+	md"""## Select Buoy
+	$(sta_b)"""
 end
 
 # ╔═╡ a1698e0e-db0d-4cd2-91b3-d530f77cd609
 begin
-    buoy=read(NOAAbuoy(),41046)
+    buoy=read(NOAAbuoy(),sta)
 	"Done with reading data"
 end
 
 # ╔═╡ c0a3ec23-2e76-40e6-bdf8-b6b774dabae5
 begin
-	var_b = @bind var Select(names(buoy.data), default="PRES")
-	md"""Select variable : $(var_b)"""
+	nms=names(buoy.data)
+	nms=nms[findall([!ismissing(buoy.data[1,nm]) for nm in nms])]
+	nm0 = (in("PRES",nms) ?  "PRES" : nms[6])
+	var_b = @bind var Select(names(buoy.data), default=nm0)
+	md"""## Select variable 
+	$(var_b)"""
 end
-
-# ╔═╡ 711e604d-ddfe-4328-a663-93dd49ff64c4
-md"""
-- `Buoy information page :` $(myurl)
-- `Variable Units.       :` $(buoy.units[var])
[…]
