### 2024-08-03 3befe59
improve plots, notebooks, new GDP_list_*csv
