### 2024-08-03 f67ad38
write/read file lists from temp file
