### 2024-08-03 6beed1e
improve docs, adding plots
