### 2024-08-04 14a2afd
update notebook deps
