+++ docs/src/visuals.md
@@ -1,24 +1,0 @@
-# Visuals
-
-The images below are from the notebook examples listed earlier in this doc.
-
-Global Data Coverage | Individual Data Platforms
-:------------------------------:|:---------------------------------:
-![](https://user-images.githubusercontent.com/20276764/208552147-d433f802-9c09-41cc-bece-f0ef424f26ea.png) | ![](https://user-images.githubusercontent.com/20276764/208441408-1ffe7508-19da-4f41-b984-58820799785a.png) 
-
-Argo Float Track            |  Argo Float Profiles 
-:------------------------------:|:---------------------------------:
-![](https://user-images.githubusercontent.com/20276764/166470235-467a9326-18ae-4934-a866-2da06ec9ec84.png)  |  ![](https://user-images.githubusercontent.com/20276764/166470217-f89d2374-f57e-4a28-8220-86179e6c1f86.png)
-
-Surface Drifter Positions | Frequency Spectra 
-:------------------------------:|:---------------------------------:
-![](https://user-images.githubusercontent.com/20276764/149673826-a43e2a44-f4e5-437b-99cb-5e032228b3af.png) | ![](https://user-images.githubusercontent.com/20276764/205257672-f8adc8fc-dea7-4dea-91dd-ab9e1c18c1c1.png)
-
-Spray Glider Data | WHOTS Mooring Data
-:------------------------------:|:---------------------------------:
-![](https://user-images.githubusercontent.com/20276764/166470390-952e89df-60ad-4a45-b015-9469c3c297de.png) | ![](https://user-images.githubusercontent.com/20276764/149675305-82364bde-e3a9-4975-8fb2-fb67e17dacc5.png)
-
-NOAA Time series | Measured Ocean Warming
-:------------------------------:|:---------------------------------:
-![](https://user-images.githubusercontent.com/20276764/166470257-8a0421ff-b147-46aa-b03b-43e5f8b4d1b3.png) | ![](https://user-images.githubusercontent.com/20276764/205256659-6505f41f-577a-481d-99e6-424073702699.png)
-
