+++ v0.2.1/examples/SatelliteAltimetry.jl
@@ -1,0 +1,2062 @@
+### A Pluto.jl notebook ###
+# v0.19.42
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ a8e0b727-a416-4aad-b660-69e5470c7e9e
+begin
+	using OceanRobots, CairoMakie, Dataverse, ArchGDAL, PlutoUI
+	OceanRobotsMakieExt=Base.get_extension(OceanRobots, :OceanRobotsMakieExt)
+	"Done with Julia packages"
+end
+
+# ╔═╡ 71e87ed3-5a9f-49aa-99af-cf144501c678
+md"""# Regional Sea Level
+
+Visualize dynamic sea level anomaly (colors) and ocean bathymetry/topography (contours) in the region of the Azores as a function of space and time. 
+
+!!! tip
+    Choose between two data sets (sources: NASA PODAAC, ESA CMEMS), select time, or generate animation.
+"""
+
+# ╔═╡ a58cc4b4-7023-4dcf-a5f4-6366be8047a3
+TableOfContents()
+
+# ╔═╡ 62e0b8a9-0025-4ce7-9538-b6114d97b762
+md"""## Visualize Data
+
+- Color shading is sea level anomaly from a gridded data product based on satellite measurement (altimetry).
+- Contours show the relief (topography, bathymetry). Light pink contours correspond to the Azores islands.
+"""
+
+# ╔═╡ c93dde18-e639-4edd-9192-f6c9eed0cb89
+@bind fil Select(["sla_podaac.nc","sla_cmems.nc"])
+
+# ╔═╡ 0c9fdfb0-bddc-4def-9954-526978491a84
+dates=OceanRobotsMakieExt.sla_dates(fil)
+
+# ╔═╡ 8ff180a4-dd71-4a70-82ab-70bc80427abb
+@bind d0 Select(dates)
+
+# ╔═╡ ff7dd5eb-5b1b-4314-9553-b8c05c4d7376
+md"""## Data Set"""
+
+# ╔═╡ 9b3c3856-9fe1-43ba-97a2-abcd5b385c1d
+sla=read(SeaLevelAnomaly(),Symbol(fil[1:end-3]))
+
+# ╔═╡ 1cf2cdb9-3c09-4b39-81cf-49318c16f531
+md"""## Apendix
+
+### Julia Codes"""
+
+# ╔═╡ 50b75406-e55f-433d-bd7b-089d975e5001
+t0=findall(dates.==d0)[1]
+
+# ╔═╡ af68228e-710c-4a5a-be48-c716592f8f45
+md"""### Data Sources
+
[…]
