### 2024-08-05 #164 @gaelforget
> The notebook [Buoy_NWP_NOAA_monthly.jl](https://juliaocean.github.io/OceanRobots.jl/stable/examples/Buoy_NWP_NOAA_monthly.html): can you add unit to the x and y axis? It it in °Fahrenheit ?

Done

> Drifter_GDP.jl:
>
> * it would be useful to define `ve` and `vn` ( zonal and meridional velocity I guess :-) )
> * the y axis is just `time` and the labels number like 300. Can you add more information about the time unit and origin (like "days since 2000-01-01" ?)

Done

> References of the PDF paper:
>
> * please check capitalization of reference 1 (e.g. co2 -> CO2,...), 3 (cmip6 - CMIP6), 6 (argo -> Argo), 11 (mitgcm -> MITgcm)... see https://stackoverflow.com/questions/2800527/preserving-all-capitalization-in-bibtex

Still to do
