### 2024-08-05 #164 @Alexander-Barth
For people not familiar with Pluto (or Julia), I think it would be nice to mention somewhere that all you need to do is to install julia and Pluto. The installation of OceanRobots.jl and other will be done automatically when you run the notebook. Is this information already somewhere?
