### 2024-09-01 cf0ba60
improve CCHDO.download + sample urls
