### 2024-09-02 3838198
more constructors, and basic plot method
