+++ docs/src/internals.md
@@ -5,6 +5,6 @@
 ```
 
 ```@autodocs
-Modules = [OceanOPS, GDP, GDP_CloudDrift, NOAA, GliderFiles, ArgoFiles, OceanSites, SLA, podaac_sla, cmems_sla]
+Modules = [OceanOPS, GDP, GDP_CloudDrift, NOAA, GliderFiles, ArgoFiles, OceanSites, CCHDO, SLA, podaac_sla, cmems_sla]
 ```
 
