+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"examples/#Notebooks","page":"Notebooks","title":"Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"note: Note\nThe static html rendering of the notebooks (this website) lack the interactivity that comes from Running Examples yourself.","category":"page"},{"location":"examples/#List-of-Notebooks","page":"Notebooks","title":"List of Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"The suite of examples includes :","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"OceanOPS.jl (➭ code link) : global ocean observing systems\nBuoy_NWP_NOAA.jl (➭ code link) : NOAA station data (a few days)\nBuoy_NWP_NOAA_monthly.jl (➭ code link) : NOAA station data (monthly means) \nMooring_WHOTS.jl (➭ code link) : WHOTS mooring data\nDrifter_GDP.jl (➭ code link) : drifter time series\nDrifter_CloudDrift.jl (➭ code link) : drifter statistics\nFloat_Argo.jl (➭ code link) : Argo profiling float data\nGlider_Spray.jl (➭ code link) : underwater glider data\nSatelliteAltimetry.jl (➭ code link) : gridded satellite data","category":"page"},{"location":"examples/#More-Notebooks","page":"Notebooks","title":"More Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"For Argo and state estimates, see ArgoData.jl\nFor drifter data simulations, see IndividualDisplacements.jl","category":"page"},{"location":"examples/#Using-Notebooks","page":"Notebooks","title":"Using Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"The examples are most easily run using Pluto.jl. To do it this way, one just needs to copy a code link provided above and paste this URL into the Pluto.jl interface.","category":"page"},{"location":"reference/#User-Interface-(API)","page":"User Interface","title":"User Interface (API)","text":"","category":"section"},{"location":"reference/","page":"User Interface","title":"User Interface","text":"Each type of ocean data gets :","category":"page"},{"location":"reference/","page":"User Interface","title":"User Interface","text":"a simple read function that downloads data if needed. \na default plot function that depicts some of the data.","category":"page"},{"location":"reference/","page":"User Interface","title":"User Interface","text":"For mapping purposes, it is useful to download country polygons.","category":"page"},{"location":"reference/","page":"User Interface","title":"User Interface","text":"using MeshArrays, Shapefile, DataDeps\npol_file=demo.download_polygons(\"ne_110m_admin_0_countries.shp\")\npol=MeshArrays.read_polygons(pol_file)\nnothing #hide","category":"page"},{"location":"reference/#Surface-Drifters","page":"User Interface","title":"Surface Drifters","text":"","category":"section"},{"location":"reference/","page":"User Interface","title":"User Inte
[…]
