+++ v0.2.2/examples/Mooring_WHOTS.jl
@@ -1,0 +1,1909 @@
+### A Pluto.jl notebook ###
+# v0.19.45
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ 5fc8cc85-7bd4-4fea-8c6f-01dabee6eb5f
+begin
+	using OceanRobots, CairoMakie, PlutoUI
+	"Done with packages"
+end
+
+# ╔═╡ 538ddc08-85bd-426b-92bc-f5dd78d38d5d
+TableOfContents()
+
+# ╔═╡ 720e1461-81b9-42b6-88be-0115a17268c0
+md"""# WHOTS mooring
+
+Here we plot time series from the [WHOTS mooring](http://www.soest.hawaii.edu/whots/wh_data.html) which measures both Ocean and Atmosphere variables. This is one of the long time series available to study Climate. See <http://www.oceansites.org/> for more data like this.
+"""
+
+# ╔═╡ e7dc453e-3bf2-4289-b55d-66eab3636d7b
+begin
+	url0="https://insideclimatenews.org/wp-content/uploads/2020/07/ohcean-mooring_noaa.jpg"
+	url1="https://aco-ssds.soest.hawaii.edu/ALOHA/img/slideshow3.jpg"
+	url2="https://www.researchgate.net/profile/Andrew-Dickson-6/publication/228352615/figure/fig1/AS:301928941670405@1448996877577/Time-series-of-atmospheric-CO-2-at-Mauna-Loa-in-ppm-mole-fraction-in-dry-air-and.png"
+	md"""
+	$(Resource(url0,:height => 100))
+	$(Resource(url1,:height => 100))
+	$(Resource(url2,:height => 100))
+
+	_Image credits: NOAA, SOEST, WHOI, Dickson._
+	"""	
+end
+
+# ╔═╡ 54768410-2886-4d72-a691-eea72aee6b67
+begin
+	y0_b = @bind y0 NumberField(2004:2018, default=2005)
+	m0_b = @bind m0 NumberField(1:12, default=1)
+	d0_b = @bind d0 NumberField(1:31, default=1)
+	y1_b = @bind y1 NumberField(2004:2018, default=2005)
+	m1_b = @bind m1 NumberField(1:12, default=2)
+	d1_b = @bind d1 NumberField(1:31, default=1)
+	md"""## Data Visualisation
+	
+	Select initial and final time here;
+	
+	- year,month,day = $(y0_b) $(m0_b) $(d0_b)
+	- year,month,day = $(y1_b) $(m1_b) $(d1_b)
+	"""
[…]
