### 2024-09-02 b3fc30d
add plotting options (salinity, chi)
