### 2024-09-02 5e519b1
add notebook, docs
