### 2024-09-03 807c80d
update notebook deps
