+++ v0.2.3/examples/Glider_Spray.jl
@@ -1,0 +1,1933 @@
+### A Pluto.jl notebook ###
+# v0.19.46
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ 0247a51e-c89b-11ec-071f-bb82fe257adc
+begin
+	using OceanRobots, PlutoUI, CairoMakie
+	"Done with Software Packages"
+end
+
+# ╔═╡ e5b9772e-cbb5-42f0-bacb-5bae6bc8d5b6
+md"""# Spray Gliders
+
+Spray gliders autonomously collect measurements in the upper kilometer of the ocean during missions typically lasting three to four months and covering 2000-3000 km through the water. ([spraydata.ucsd.edu](http://spraydata.ucsd.edu)).
+"""
+
+# ╔═╡ 0a016d12-de41-45f5-800d-af8f953720e4
+begin
+	url1="https://timesofsandiego.com/wp-content/uploads/2014/06/Spray-Glider.jpg"
+	url2="https://www.whoi.edu/wp-content/uploads/2019/01/GP020002-1-14_506075_510753-1.jpg"
+	url3="https://spraydata.ucsd.edu/assets/img/GS_KE_for_web.png"
+	url4="https://spraydata.ucsd.edu/products/cugn-climatology/plots/66/meanmap/66_meanmap_xz_geov.png"
+	md"""
+	$(Resource(url2,:height => 100))
+	$(Resource(url1,:height => 100))
+	$(Resource(url4,:height => 100))
+	$(Resource(url3,:height => 100))
+
+	_Image Credits : WHOI, UCSD._
+	"""
+end
+
+# ╔═╡ d9ffd0a8-85e3-4e2a-9d6a-81f3dbd3ee31
+TableOfContents()
+
+# ╔═╡ b8db58ca-ed7a-4174-a6b0-e53925b5887a
+begin
+	MID_bind = @bind MID Select(["CUGN_along.nc", "GulfStream.nc"], default="GulfStream.nc")
+	md"""## Interactive Visualization
+	
+	- Select data set : $(MID_bind)"""
+end
+
+
+# ╔═╡ 68028522-205e-4b41-b3c0-2e3b09c2d8a7
+md"""## Appendix
+
+### Julia Packages
+"""
+
+# ╔═╡ e729416a-a2b2-46ce-bb1e-ededeb8c75b3
+md"""### Data Ingestion"""
+
+# ╔═╡ 9523dc0d-1758-4e0f-864c-4ab253bf11a9
+begin
+    gliders=read(Gliders(),MID)
+    ng=length(unique(gliders.data.ID))
+	ID_bind = @bind ID NumberField(1:ng, default=1)
+	"Done with Data Ingestion"
[…]
