### 2024-09-03 f23b92e
streamline handling of file names in CCHDO
