+++ src/types.jl
@@ -74,12 +74,10 @@ SeaLevelAnomaly() = SeaLevelAnomaly(:unknown,[],"")
 Base.@kwdef struct OceanExpedition <: AbstractOceanRobotData
     ID::String
     data::Union{Array,Dataset}
-    list_files::DataFrame
+    path::String
 end
 
 OceanExpedition(ID::String)=begin
     path=CCHDO.download(ID)
-    filenames=joinpath.(Ref(path),readdir(path))
-    list1=DataFrames.DataFrame(:filename=>filenames)
-    OceanExpedition(ID,[],list1)
+    OceanExpedition(ID,[],path)
 end
