### 2024-09-03 cc89262
rename data structure as ShipCruise
