+++ examples/ShipCruise_CCHDO.jl
@@ -4,37 +4,104 @@
 using Markdown
 using InteractiveUtils
 
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
 # ╔═╡ e741dbc2-695b-11ef-1962-5769d79825c0
-using OceanRobots, CairoMakie
+using OceanRobots, CairoMakie, PlutoUI
 
 # ╔═╡ c90d4b04-9810-4736-8f73-bb6ee2c42577
 md"""# CTD profiles from ocean ship expeditions
 
+CCHDO provides access to high quality, global, vessel-based CTD and hydrographic data from GO-SHIP, WOCE, CLIVAR and other repeat hydrography programs. 
+
 !!! note
-    Data displayed here were collected by research vessells. For more information see <https://cchdo.ucsd.edu>
+    For more information see <https://cchdo.ucsd.edu>
 """
 
-# ╔═╡ eab5b722-be6b-4341-b870-5a98cf6a32c3
-#plot(cruise,variable="chi_up",colorrange=(-12,-10))
-
-# ╔═╡ 17b1bef9-719a-4668-a067-c8dabdab9dc2
-md"""## Data and Software"""
+# ╔═╡ 4a1784d6-7850-4c05-b302-013db8966c3a
+@bind ID Select(["33RR20160208","74EQ20151206","33RO20131223"])
 
 # ╔═╡ f5415f1d-0e98-41e3-a6fb-459580141611
-cruise=OceanExpedition("33RR20160208")
+cruise=ShipCruise(ID)
 
 # ╔═╡ 19a5cf7c-fa71-435b-bdff-d369658d3260
 plot(cruise,variable="salinity",colorrange=(33.5,35.0))
+
+# ╔═╡ a7186ef8-f1de-437f-99eb-fe6553da49b5
+md"""### Chipod data
+
+Some of these ocean expeditions measure ocean mixing using `chipod` sensors. 
+
+!!! note
+    For more information see <http://microstructure.ucsd.edu/>
+"""
+
+# ╔═╡ 457d85eb-08a8-4304-82a7-0372514b1bec
+@bind va Select(["chi_up","chi_dn","KT_up","KT_dn"])
+
+# ╔═╡ beed86ef-5e86-45a4-8842-f21f2dcef504
+ds=CCHDO.open_chipod_file(cruise)
+
+# ╔═╡ 17b1bef9-719a-4668-a067-c8dabdab9dc2
+md"""## Julia Code"""
+
+# ╔═╡ bc007619-ff00-49c8-a8e4-e2e8a96a0326
+function chipod_plot(x;variable="chi_up",colorrange=(-12.0,-10.0),apply_log10=true,markersize=3)
+	
[…]
