### 2024-09-03 556cb11
improve plotting for ShipCruise
