+++ dev/examples/OceanExpedition_CCHDO.jl
@@ -1,1792 +1,0 @@
-### A Pluto.jl notebook ###
-# v0.19.46
-
-using Markdown
-using InteractiveUtils
-
-# ╔═╡ e741dbc2-695b-11ef-1962-5769d79825c0
-using OceanRobots, CairoMakie
-
-# ╔═╡ c90d4b04-9810-4736-8f73-bb6ee2c42577
-md"""# CTD profiles from ocean ship expeditions
-
-!!! note
-    Data displayed here were collected by research vessells. For more information see <https://cchdo.ucsd.edu>
-"""
-
-# ╔═╡ eab5b722-be6b-4341-b870-5a98cf6a32c3
-#plot(cruise,variable="chi_up",colorrange=(-12,-10))
-
-# ╔═╡ 17b1bef9-719a-4668-a067-c8dabdab9dc2
-md"""## Data and Software"""
-
-# ╔═╡ f5415f1d-0e98-41e3-a6fb-459580141611
-cruise=OceanExpedition("33RR20160208")
-
-# ╔═╡ 19a5cf7c-fa71-435b-bdff-d369658d3260
-plot(cruise,variable="salinity",colorrange=(33.5,35.0))
-
-# ╔═╡ 00000000-0000-0000-0000-000000000001
-PLUTO_PROJECT_TOML_CONTENTS = """
-[deps]
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
-OceanRobots = "0b51df41-3294-4961-8d23-db645e32016d"
-
-[compat]
-CairoMakie = "~0.12.9"
-OceanRobots = "~0.2.1"
-"""
-
-# ╔═╡ 00000000-0000-0000-0000-000000000002
-PLUTO_MANIFEST_TOML_CONTENTS = """
-# This file is machine-generated - editing it directly is not advised
-
-julia_version = "1.10.4"
-manifest_format = "2.0"
-project_hash = "0bb01b4475c900c10989d79fa18fb64c96d81b58"
-
-[[deps.AbstractFFTs]]
-deps = ["LinearAlgebra"]
-git-tree-sha1 = "d92ad398961a3ed262d8bf04a1a2b8340f915fef"
-uuid = "621f4979-c628-5d54-868e-fcf4e3e8185c"
-version = "1.5.0"
-weakdeps = ["ChainRulesCore", "Test"]
-
-    [deps.AbstractFFTs.extensions]
-    AbstractFFTsChainRulesCoreExt = "ChainRulesCore"
-    AbstractFFTsTestExt = "Test"
-
-[[deps.AbstractTrees]]
-git-tree-sha1 = "2d9c9a55f9c93e8887ad391fbae72f8ef55e1177"
-uuid = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
-version = "0.4.5"
-
-[[deps.Adapt]]
-deps = ["LinearAlgebra", "Requires"]
-git-tree-sha1 = "6a55b747d1812e699320963ffde36f1ebdda4099"
-uuid = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-version = "4.0.4"
-weakdeps = ["StaticArrays"]
-
-    [deps.Adapt.extensions]
-    AdaptStaticArraysExt = "StaticArrays"
-
-[[deps.AdaptivePredicates]]
-git-tree-sha1 = "7d5da5dd472490d048b081ca1bda4a7821b06456"
-uuid = "35492f91-a3bd-45ad-95db-fcad7dcfedb7"
-version = "1.1.1"
-
[…]
