+++ v0.2.4/examples/Float_Argo.jl
@@ -1,0 +1,2107 @@
+### A Pluto.jl notebook ###
+# v0.19.46
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ 8539e67f-6361-409a-9dbe-e8dcb2c7e10d
+begin
+	using OceanRobots, ArgoData, CairoMakie, PlutoUI, PrettyTables, MeshArrays
+	"Done with packages"
+end
+
+# ╔═╡ dc017980-cde3-49bc-a5e0-44be9bb08cec
+begin
+	using Shapefile, DataDeps
+	pol_file=demo.download_polygons("ne_110m_admin_0_countries.shp")
+	pol=MeshArrays.read_polygons(pol_file)
+	"Done with country boundaries"
+end
+
+# ╔═╡ 89647c4e-e568-4d3a-af57-bcc541744e38
+md"""# Argo Floats
+
+The standard Argo platform is a profiling float that drifts at 1000m depth. They come back to the surface every ten days, just after collecting vertocal profiles of temperature and salinity, and transmit their data via satellite. 
+
+!!! tip
+    For more information about the international Argo program please refer to, for example, <https://argo.ucsd.edu> or <https://biogeochemical-argo.org>.
+"""
+
+# ╔═╡ 09a2aa6e-16bc-4582-9d5f-c50432e3f0ca
+begin
+	url1="https://www.researchgate.net/profile/Taiyo-Kobayashi/publication/236890114/figure/fig1/AS:299396928425984@1448393198754/Distribution-of-profiling-floats-under-Argo-by-country-Total-number-of-floats-is-3256-as.png"
+	url2="https://www.researchgate.net/profile/Taiyo-Kobayashi/publication/236890114/figure/fig2/AS:299396928425984@1448393198754/Distribution-of-profiling-floats-under-Argo-by-country-Total-number-of-floats-is-3256-as.png"
+	url4="https://floats.pmel.noaa.gov/sites/default/files/photos/Deep%20Argo/15.%20Deep%20SOLO%20float%20starts%20its%20mission%20by%20sinking%20after%20deployment%20during%20HOT%20Cruise%20302.jpg"
+
+	md"""
+
+	
+	$(Resource(url1,:height => 150)) 
+	$(Resource(url2,:height => 150))
+	$(Resource(url4,:height => 150))
+
+	_Picture credits : Taiyo Kobayashi, IFREMER, NOAA._
+	"""
[…]
