+++ v0.2.4/examples/Drifter_CloudDrift.jl
@@ -1,0 +1,1896 @@
+### A Pluto.jl notebook ###
+# v0.19.46
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ bf29d9b6-be57-11ec-2b83-87ce1f770075
+using CairoMakie, OceanRobots, FFTW, PlutoUI
+
+# ╔═╡ 97537903-fb3b-4837-9e7c-461cb9e4ef5e
+md"""# Transpose _Cloud-Drift_ workflow into Julia
+
+See <https://github.com/Cloud-Drift/earthcube-meeting-2022> for context and detail.
+"""
+
+# ╔═╡ c56e7f24-7be8-4209-8fc9-843cb674fd6e
+TableOfContents()
+
+# ╔═╡ f3c351c7-3fba-4ee8-8522-80fd0e8921af
+md"""## Visualize Data"""
+
+# ╔═╡ 05b32012-927d-42b4-8a18-fdfcb051a3fd
+md"""## Per Trajectory Statistics"""
+
+
+# ╔═╡ 89c82c13-9221-48b8-9138-79b91603d7e6
+md"""## Appendix
+
+Packages, grid, functions used in the notebook are defined here.
+"""
+
+# ╔═╡ d0362426-0609-4fcf-a888-77892db9b7e7
+begin
+    GDP_CD=read(CloudDrift(),"")
+    df=GDP_CD.data.main
+    df_subset=GDP_CD.data.subset
+	"Done reading data"
+end
+
+# ╔═╡ 60479536-8238-4077-bdcd-776e07330621
+let
+	ii=findall(isfinite.(df_subset.longitude.*df_subset.latitude))	
+	scatter(df_subset.longitude[ii], df_subset.latitude[ii], markersize=2,
+		color=sqrt.(df_subset.ve.^2+df_subset.vn.^2), colorrange=(0,2.5), colormap=:speed)
+end
+
+# ╔═╡ d4af67e7-2308-4135-929e-c7fc6572d6c7
+heatmap(GDP_CD.data.grid.lon,GDP_CD.data.grid.lat,GDP_CD.data.ve)
+
+# ╔═╡ a25027b5-5b0e-4190-b4b6-987f4e16ccd8
+begin
+	gdf2=GDP_CloudDrift.groupby(df,:ID)
+	df2=GDP_CloudDrift.trajectory_stats(gdf2)
+	trajectory_periodogram(cv) = 1/24 *abs.(fft(cv[:])).^2
+	periodograms=[trajectory_periodogram(tmp.cv[:]) for tmp in gdf2]
+	df2[1,:]
+end
+
+# ╔═╡ efc6b8b3-a55b-4830-953c-2a51ce162604
+begin
+	# ( for detail, see original notebook @ https://github.com/Cloud-Drift/earthcube-meeting-2022 )
+
+    ii=1
+
+    # calculate inertial frequency in cycle per day [cpd] from the mean latitude of trajectory
+    omega = 7.2921159e-5  # Earth's rotation rate [rad/s]
+    seconds_per_day = 60*60*24  # [s]
+    fi = -2*omega*(seconds_per_day/(2*pi))*sind(df2[ii,:latitude])
+
+    # define frequency scale/abscissa
+    dt = 1/24  # [-]
+    f = fftfreq(length(periodograms[ii]), dt) /dt/dt
+    ss = log10.(periodograms[ii])  # spectrum of the ith trajectory 
+
+    # plot
+    function plot1(f,ss)
[…]
