### 2024-09-03 c12540b
update & improve CCHDO notebook
