+++ examples/ShipCruise_CCHDO.jl
@@ -18,12 +18,15 @@ end
 using OceanRobots, CairoMakie, PlutoUI
 
 # ╔═╡ c90d4b04-9810-4736-8f73-bb6ee2c42577
-md"""# CTD profiles from ocean ship expeditions
+md"""# CTD profiles from oceanographic expeditions
 
-CCHDO provides access to high quality, global, vessel-based CTD and hydrographic data from GO-SHIP, WOCE, CLIVAR and other repeat hydrography programs. 
+CCHDO provides access to high quality, global, ship-based CTD (Conductivity-Temperature-Depth) and hydrographic data from GO-SHIP, WOCE, CLIVAR and other repeat hydrography programs. 
 
 !!! note
     For more information see <https://cchdo.ucsd.edu>
+
+### Select Cruise
+
 """
 
 # ╔═╡ 4a1784d6-7850-4c05-b302-013db8966c3a
@@ -32,8 +35,29 @@ CCHDO provides access to high quality, global, vessel-based CTD and hydrographic
 # ╔═╡ f5415f1d-0e98-41e3-a6fb-459580141611
 cruise=ShipCruise(ID)
 
+# ╔═╡ fa2c9be1-1198-4852-b7b1-623e42f5c871
+begin
+	vari_bind = @bind vari Select(["temperature","salinity"],default="salinity")
+	cmin_bind = @bind cmin TextField(default="33.5")
+	cmax_bind = @bind cmax TextField(default="35.0")
+	md"""
+!!! note
+	Selecting a new data set will start by downloading it, if needed, which may take a little while.
+
+### Plot One Variable
+
+- variable : $(vari_bind)
+- color range minimum	: $(cmin_bind)
+- color range maximum	: $(cmax_bind)
+	"""
+end
+	
+
 # ╔═╡ 19a5cf7c-fa71-435b-bdff-d369658d3260
-plot(cruise,variable="salinity",colorrange=(33.5,35.0))
+begin
+	crng=(parse(Float64,cmin),parse(Float64,cmax))
+	plot(cruise,variable=vari,colorrange=crng)
+end
 
 # ╔═╡ a7186ef8-f1de-437f-99eb-fe6553da49b5
 md"""### Chipod data
@@ -101,7 +125,6 @@ PlutoUI = "7f904dfe-b85e-4ff6-b463-dae2292396a8"
 
 [compat]
 CairoMakie = "~0.12.9"
-OceanRobots = "~0.2.1"
 PlutoUI = "~0.7.60"
 """
 
@@ -111,7 +134,7 @@ PLUTO_MANIFEST_TOML_CONTENTS = """
 
 julia_version = "1.10.4"
 manifest_format = "2.0"
-project_hash = "8e1da76a5f003c9f989dfb30f8415f5f369f48d5"
+project_hash = "43a93e7fb194aa665ccd472bb18457ac82ed7c8a"
 
 [[deps.AbstractFFTs]]
 deps = ["LinearAlgebra"]
@@ -642,10 +665,10 @@ uuid = "42e2da0e-8278-4e71-bc24-59509adca0fe"
 version = "1.0.2"
 
 [[deps.HDF5_jll]]
[…]
