+++ dev/examples/ShipCruise_CCHDO.jl
@@ -4,37 +4,128 @@
 using Markdown
 using InteractiveUtils
 
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
 # ╔═╡ e741dbc2-695b-11ef-1962-5769d79825c0
-using OceanRobots, CairoMakie
+using OceanRobots, CairoMakie, PlutoUI
 
 # ╔═╡ c90d4b04-9810-4736-8f73-bb6ee2c42577
-md"""# CTD profiles from ocean ship expeditions
+md"""# CTD profiles from oceanographic expeditions
+
+CCHDO provides access to high quality, global, ship-based CTD (Conductivity-Temperature-Depth) and hydrographic data from GO-SHIP, WOCE, CLIVAR and other repeat hydrography programs. 
 
 !!! note
-    Data displayed here were collected by research vessells. For more information see <https://cchdo.ucsd.edu>
+    For more information see <https://cchdo.ucsd.edu>
+
+### Select Cruise
+
 """
 
-# ╔═╡ eab5b722-be6b-4341-b870-5a98cf6a32c3
-#plot(cruise,variable="chi_up",colorrange=(-12,-10))
-
-# ╔═╡ 17b1bef9-719a-4668-a067-c8dabdab9dc2
-md"""## Data and Software"""
+# ╔═╡ 4a1784d6-7850-4c05-b302-013db8966c3a
+@bind ID Select(["33RR20160208","74EQ20151206","33RO20131223"])
 
 # ╔═╡ f5415f1d-0e98-41e3-a6fb-459580141611
-cruise=OceanExpedition("33RR20160208")
+cruise=ShipCruise(ID)
+
+# ╔═╡ fa2c9be1-1198-4852-b7b1-623e42f5c871
+begin
+	vari_bind = @bind vari Select(["temperature","salinity"],default="salinity")
+	cmin_bind = @bind cmin TextField(default="33.5")
+	cmax_bind = @bind cmax TextField(default="35.0")
+	md"""
+!!! note
+	Selecting a new data set will start by downloading it, if needed, which may take a little while.
+
+### Plot One Variable
+
+- variable : $(vari_bind)
+- color range minimum	: $(cmin_bind)
+- color range maximum	: $(cmax_bind)
+	"""
+end
+	
 
 # ╔═╡ 19a5cf7c-fa71-435b-bdff-d369658d3260
-plot(cruise,variable="salinity",colorrange=(33.5,35.0))
+begin
+	crng=(parse(Float64,cmin),parse(Float64,cmax))
[…]
