### 2024-09-05 7a92280
add user directions to run notebooks
