+++ docs/Project.toml
@@ -3,6 +3,8 @@ ArgoData = "9eb831cf-c491-48dc-bed4-6aca718df73c"
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Git = "d7ba0133-e1db-5d97-8f8c-041e4b3a1eb2"
 MeshArrays = "cb8c808f-1acf-59a3-9d2b-6e38d009f683"
+Pluto = "c3e4b0f8-55cb-11ea-2926-15256bba5781"
 PlutoSliderServer = "2fc8631c-6f24-4c5b-bca7-cbb509c42db4"
 Shapefile = "8e980c4a-a4fe-5da2-b3a7-4b4b0353a2f4"
