### 2024-09-05 9369e59
add user directions to run notebooks (part 2)
