+++ docs/src/reference.md
@@ -7,10 +7,7 @@ Each type of ocean data gets :
 
 # Supported Datasets
 
-!!! note
-    For mapping purposes, it is useful to download country polygons.
-
-```@example ex1
+```@setup ex1
 using MeshArrays, Shapefile, DataDeps
 pol_file=demo.download_polygons("ne_110m_admin_0_countries.shp")
 pol=MeshArrays.read_polygons(pol_file)
@@ -75,7 +72,15 @@ gliders=read(Gliders(),"GulfStream.nc")
 plot(gliders,1,pol=pol)
 ```
 
-## Sea Level Anomaly
+!!! note
+    To put data in context, it is useful to download country polygons or gridded data sets.
+
+```@example ex1
+using MeshArrays, Shapefile, DataDeps
+pol_file=demo.download_polygons("ne_110m_admin_0_countries.shp")
+pol=MeshArrays.read_polygons(pol_file)
+nothing #hide
+```
 
 ```@example ex1
 using OceanRobots, CairoMakie
