### 2024-09-05 #164 @gaelforget
> For people not familiar with Pluto (or Julia), I think it would be nice to mention somewhere that all you need to do is to install julia and Pluto. The installation of OceanRobots.jl and other will be done automatically when you run the notebook. Is this information already somewhere?

Good point! Thanks for the suggestion.

I added on this in the new How-To section , https://juliaocean.github.io/OceanRobots.jl/dev/examples/#How-To
