### 2024-09-05 #164 @gaelforget
> Concerning "Does installation proceed as outlined in the documentation?": I did not see any specific instruction (or did I miss them?), but I can confirm that package can be installed as any julia package.

I added installation docs in the new How-To section , https://juliaocean.github.io/OceanRobots.jl/dev/examples/#How-To
