+++ paper/header.tex
@@ -1,19 +1,16 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{My JuliaCon proceeding}
+\title{Digital Twins for Ocean Robots}
 
-\author[1]{1st author}
-\author[1, 2]{2nd author}
-\author[2]{3rd author}
-\affil[1]{University}
-\affil[2]{National Lab}
+\author[1]{Gaël Forget}
+\affil[1]{Earth, Atmospheric and Planetary Sciences, Massachusetts Institute of Technology, 77 Massachusetts Avenue, Cambridge, 02139, MA, USA}
 
-\keywords{Julia, Optimization, Game theory, Compiler}
+\keywords{Julia, OptimizatOcean, Climate, Robot, Observation, Platform, Sensor, Drifter, Buoy, Profiler, Satellite, Modeling, Artificial Intelligence, Data Assimilation, Parameter Inference, Climatology, Geospatial, Statistics}
 
 \hypersetup{
-pdftitle = {My JuliaCon proceeding},
+pdftitle = {Digital Twins for Ocean Robots},
 pdfsubject = {JuliaCon 2022 Proceedings},
-pdfauthor = {1st author, 2nd author, 3rd author},
-pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+pdfauthor = {Gaël Forget},
+pdfkeywords = {Julia, OptimizatOcean, Climate, Robot, Observation, Platform, Sensor, Drifter, Buoy, Profiler, Satellite, Modeling, Artificial Intelligence, Data Assimilation, Parameter Inference, Climatology, Geospatial, Statistics},
 }
 
