+++ paper/ref.bib
@@ -1,7 +1,7 @@
 
-@ARTICLE{Bittig2018,
+@article{Bittig2018,
 AUTHOR={Bittig, Henry C. and Steinhoff, Tobias and Claustre, Hervé and Fiedler, Björn and Williams, Nancy L. and Sauzède, Raphaëlle and Körtzinger, Arne and Gattuso, Jean-Pierre},   
-TITLE={An Alternative to Static Climatologies: Robust Estimation of Open Ocean CO2 Variables and Nutrient Concentrations From T, S, and O2 Data Using Bayesian Neural Networks},      
+TITLE={An Alternative to Static Climatologies: Robust Estimation of Open Ocean {CO2} Variables and Nutrient Concentrations From {T, S, and O2} Data Using Bayesian Neural Networks},      
 JOURNAL={Frontiers in Marine Science},     
 VOLUME={5},          
 YEAR={2018},      
@@ -10,7 +10,7 @@ DOI={10.3389/fmars.2018.00328},
 	ISSN={2296-7745},   
 }
 
-@ARTICLE{Sauzède2017,
+@article{Sauzède2017,
   AUTHOR={Sauzède, Raphaëlle and Bittig, Henry C. and Claustre, Hervé and Pasqueron de Fommervault, Orens and Gattuso, Jean-Pierre and Legendre, Louis and Johnson, Kenneth S.},    
 TITLE={Estimates of Water-Column Nutrient Concentrations and Carbonate System Parameters in the Global Ocean: A Novel Approach Based on Neural Networks},      
 JOURNAL={Frontiers in Marine Science},      
@@ -85,7 +85,7 @@ ISSN={2296-7745},
   journal = {Journal of Open Source Software}
 }
 
-@Article{Hector2015,
+@article{Hector2015,
 AUTHOR = {Hartin, C. A. and Patel, P. and Schwarber, A. and Link, R. P. and Bond-Lamberty, B. P.},
 TITLE = {A simple object-oriented and open-source model for scientific and policy analyses of the global climate system – Hector v1.0},
 JOURNAL = {Geoscientific Model Development},
@@ -97,9 +97,9 @@ URL = {https://gmd.copernicus.org/articles/8/939/2015/},
 DOI = {10.5194/gmd-8-939-2015}
 }
 
-@Article{CMIP6,
+@article{CMIP6,
 AUTHOR = {Eyring, V. and Bony, S. and Meehl, G. A. and Senior, C. A. and Stevens, B. and Stouffer, R. J. and Taylor, K. E.},
-TITLE = {Overview of the Coupled Model Intercomparison Project Phase 6 (CMIP6)
+TITLE = {Overview of the Coupled Model Intercomparison Project Phase 6 ({CMIP6})
 experimental design and organization},
 JOURNAL = {Geoscientific Model Development},
 VOLUME = {9},
@@ -111,9 +111,9 @@ DOI = {10.5194/gmd-9-1937-2016}
 }
 
 
-@ARTICLE{Roemmich2019,
+@article{Roemmich2019,
[…]
