+++ .github/workflows/ci.yml
@@ -39,7 +39,7 @@ jobs:
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
-      - uses: julia-actions/julia-processcoverage@v2
+      - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v4
         env:
           token: ${{ secrets.CODECOV_TOKEN }}
