+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"examples/#Tutorial-Notebooks","page":"Notebooks","title":"Tutorial Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"note: Note\nThe version of the notebooks found in the online docs is static html rendering (this website).\nThe version found in the src folder is the underlying code of the notebooks (.jl files).\nTo run the notebooks interactively see How-To section.","category":"page"},{"location":"examples/#Included-Notebooks","page":"Notebooks","title":"Included Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"OceanOPS.jl (➭ code link) : global ocean observing systems\nShipCruise_CCHDO.jl (➭ code link) : ship CTD and other data\nFloat_Argo.jl (➭ code link) : Argo profiling float data\nDrifter_GDP.jl (➭ code link) : drifter time series\nDrifter_CloudDrift.jl (➭ code link) : drifter statistics\nGlider_Spray.jl (➭ code link) : underwater glider data\nBuoy_NWP_NOAA.jl (➭ code link) : NOAA station data (a few days)\nBuoy_NWP_NOAA_monthly.jl (➭ code link) : NOAA station data (monthly means) \nMooring_WHOTS.jl (➭ code link) : WHOTS mooring data","category":"page"},{"location":"examples/#More-Notebooks","page":"Notebooks","title":"More Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"For Argo and state estimates, see ArgoData.jl\nFor simulations of drifter data, see IndividualDisplacements.jl\nSatelliteAltimetry.jl (➭ code link) : gridded satellite data","category":"page"},{"location":"examples/#How-To","page":"Notebooks","title":"How-To","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"To install OceanRobots.jl in julia proceed as usual via the package manager (using Pkg; Pkg.add(OceanRobots)).","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"To run a notebook interactively (.jl files) you want to use Pluto.jl. For example, copy and paste one of the above code links in the Pluto.jl interface. This will let you spin up the notebook in a web browser from the copied URL.","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"All you need to do beforehand is to install julia and Pluto.jl. The installation of OceanRobots.jl and other Julia packages will then happen automatically when you run the notebook. ","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"You can also download the notebooks and run them as normal Julia programs. We recommend runing each notebook in its own environment as shown below. ","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"using Pkg; Pkg.add(\"Pluto\"); using Pluto\n\nnotebook=\"MITgcm.jl/examples/Float_Argo.jl\"\nimport OceanRobots; path=dirname(dirname(pathof(OceanRobots))) #hide\nnotebook=joinpath(path,\"examples\",\"Float_Argo.jl\"
[…]
