+++ v0.2.5/examples/Drifter_GDP.jl
@@ -1,0 +1,1910 @@
+### A Pluto.jl notebook ###
+# v0.19.46
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ d61376ca-76f2-11ec-0393-83d9b38831ca
+begin
+	using OceanRobots, PlutoUI, CairoMakie
+	"Done with packages"
+end
+
+# ╔═╡ e6d47534-3bdb-401f-8758-12ebce4232f6
+TableOfContents()
+
+# ╔═╡ 84964782-cdd1-4773-a134-04d486e69856
+md"""# Surface Drifters
+
+Here we learn to download, read, and plot one hourly drifter data file from [NOAA webpage](https://www.aoml.noaa.gov/ftp/pub/phod/lumpkin/hourly/v2.00/netcdf/) or its [ftp server](ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/). 
+
+!!! note
+    See [global drifter website](https://www.aoml.noaa.gov/phod/gdp/hourly_data.php) for more information about this data set.
+"""
+
+# ╔═╡ 568d54ae-19a6-4ff4-ae0b-e131693029e5
+begin
+	url0="https://www.aoml.noaa.gov/wp-content/uploads/2020/12/GDP-map.jpg"
+	url1="https://www.aoml.noaa.gov/wp-content/uploads/2020/07/Global-Drifter-Program-300x300.jpg"
+	url2="https://www.hidrografico.pt/recursos/imagens/noticias/2021/01/20210108-NOAA-GDP-1.jpg"
+	url3="https://www.aoml.noaa.gov/wp-content/uploads/2020/08/GlobalDrifterProgram.jpg"
+	
+	#	http://www.aoml.noaa.gov/phod/gdp/interactive/drifter_array.html"
+	md"""	
+ 	$(PlutoUI.Resource(url0,:height => 100))
+ 	$(PlutoUI.Resource(url1,:height => 100))
+	$(PlutoUI.Resource(url2,:height => 100))
+ 	$(PlutoUI.Resource(url3,:height => 100))
+	"""
+end
+
+# ╔═╡ 4ada19d2-b4ce-463c-b771-f067b2982db0
+md"""## Interactive Display"""
+
+# ╔═╡ 760bc397-cdd3-49eb-80ef-7e4bc73885ce
+md"""## Appendix"""
+
+# ╔═╡ 37186681-41f5-4bbc-b8f6-fde37c7b5130
+md"""### Julia"""
+
+# ╔═╡ 156ff1f1-ca05-4ac6-8351-69b6c19ed8fa
[…]
