+++ v0.2.5/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"examples/#Tutorial-Notebooks","page":"Notebooks","title":"Tutorial Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"note: Note\nThe version of the notebooks found in the online docs is static html rendering (this website).\nThe version found in the src folder is the underlying code of the notebooks (.jl files).\nTo run the notebooks interactively see How-To section.","category":"page"},{"location":"examples/#Included-Notebooks","page":"Notebooks","title":"Included Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"OceanOPS.html (➭ code link) : global ocean observing systems\nShipCruise_CCHDO.html (➭ code link) : ship CTD and other data\nFloat_Argo.html (➭ code link) : Argo profiling float data\nDrifter_GDP.html (➭ code link) : drifter time series\nDrifter_CloudDrift.html (➭ code link) : drifter statistics\nGlider_Spray.html (➭ code link) : underwater glider data\nBuoy_NWP_NOAA.html (➭ code link) : NOAA station data (a few days)\nBuoy_NWP_NOAA_monthly.html (➭ code link) : NOAA station data (monthly means) \nMooring_WHOTS.html (➭ code link) : WHOTS mooring data","category":"page"},{"location":"examples/#More-Notebooks","page":"Notebooks","title":"More Notebooks","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"For Argo and state estimates, see ArgoData.jl\nFor simulations of drifter data, see IndividualDisplacements.jl\nSatelliteAltimetry.html (➭ code link) : gridded satellite data","category":"page"},{"location":"examples/#How-To","page":"Notebooks","title":"How-To","text":"","category":"section"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"To install OceanRobots.jl in julia proceed as usual via the package manager (using Pkg; Pkg.add(OceanRobots)).","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"To run a notebook interactively (.jl files) you want to use Pluto.jl. For example, copy and paste one of the above code links in the Pluto.jl interface. This will let you spin up the notebook in a web browser from the copied URL.","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"All you need to do beforehand is to install julia and Pluto.jl. The installation of OceanRobots.jl and other Julia packages will then happen automatically when you run the notebook. ","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"You can also download the notebooks folder and run them as normal Julia programs. We recommend runing each notebook in its own environment as shown below. ","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"note: Note\nTo download OceanRobots.jl folder, which includes the notebooks folder, you can use Git.jl.","category":"page"},{"location":"examples/","page":"Notebooks","title":"Notebooks","text":"usi
[…]
