+++ v0.2.5/examples/ShipCruise_CCHDO.jl
@@ -1,0 +1,1931 @@
+### A Pluto.jl notebook ###
+# v0.19.46
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ e741dbc2-695b-11ef-1962-5769d79825c0
+using OceanRobots, CairoMakie, PlutoUI
+
+# ╔═╡ c90d4b04-9810-4736-8f73-bb6ee2c42577
+md"""# CTD profiles from oceanographic expeditions
+
+CCHDO provides access to high quality, global, ship-based CTD (Conductivity-Temperature-Depth) and hydrographic data from GO-SHIP, WOCE, CLIVAR and other repeat hydrography programs. 
+
+!!! note
+    For more information see <https://cchdo.ucsd.edu>
+
+### Select Cruise
+
+"""
+
+# ╔═╡ 4a1784d6-7850-4c05-b302-013db8966c3a
+@bind ID Select(["33RR20160208","74EQ20151206","33RO20131223"])
+
+# ╔═╡ f5415f1d-0e98-41e3-a6fb-459580141611
+cruise=ShipCruise(ID)
+
+# ╔═╡ fa2c9be1-1198-4852-b7b1-623e42f5c871
+begin
+	vari_bind = @bind vari Select(["temperature","salinity"],default="salinity")
+	cmin_bind = @bind cmin TextField(default="33.5")
+	cmax_bind = @bind cmax TextField(default="35.0")
+	md"""
+!!! note
+	Selecting a new data set will start by downloading it, if needed, which may take a little while.
+
+### Plot One Variable
+
+- variable : $(vari_bind)
+- color range minimum	: $(cmin_bind)
+- color range maximum	: $(cmax_bind)
+	"""
+end
+	
+
+# ╔═╡ 19a5cf7c-fa71-435b-bdff-d369658d3260
+begin
+	crng=(parse(Float64,cmin),parse(Float64,cmax))
+	plot(cruise,variable=vari,colorrange=crng)
+end
+
+# ╔═╡ a7186ef8-f1de-437f-99eb-fe6553da49b5
+md"""### Chipod data
+
+Some of these ocean expeditions measure ocean mixing using `chipod` sensors. 
+
+!!! note
+    For more information see <http://microstructure.ucsd.edu/>
+"""
+
+# ╔═╡ 457d85eb-08a8-4304-82a7-0372514b1bec
+@bind va Select(["chi_up","chi_dn","KT_up","KT_dn"])
+
[…]

[…]
