### 2024-09-07 66dd8d1
update notebook deps
