+++ v0.2.6/examples/Buoy_NWP_NOAA.jl
@@ -1,0 +1,1964 @@
+### A Pluto.jl notebook ###
+# v0.19.46
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+end
+
+# ╔═╡ c8d36a50-0473-410a-91fa-102f3d071388
+begin
+	using Downloads, ClimateModels
+	using OceanRobots, PlutoUI, CairoMakie
+	"Done with packages"
+end
+
+# ╔═╡ d6feeecc-87db-4bfd-8341-adede2df8ea0
+begin
+	img_url="https://nosc.noaa.gov/OSC/images/Kukui_and_NOAA_buoy_1.177308.JPG"
+	md"""# NOAA Buoys
+
+	Visualize real time data for buoys that report `Standard Meteorological Data`.
+	
+	!!! note 
+	    For more information about NOAA Buoys, visit <https://www.ndbc.noaa.gov/> and links below.
+
+	$(Resource(img_url,:width => 250))
+	"""
+end
+
+# ╔═╡ ee3373bb-a848-4919-9163-384b6f33632a
+TableOfContents()
+
+# ╔═╡ 46ff94e2-1b1e-454d-ab92-89be94123a47
+md"""## Code and Data
+"""
+
+# ╔═╡ db662fb4-7413-11ec-1af6-43b18c0c15a9
+begin
+	#See https://www.ndbc.noaa.gov/
+    stations=[41046, 44065,  
+		31005, 41047, 41117, 42058, 44027, 44084, 45002, 45136, 45164,
+		45206, 46028, 46077, 46146, 46237, 46277, 51208, 62121, 64045]	
+	NOAA.download(stations)
+   	"Done with downloading data"
+end
+
+# ╔═╡ 04354f30-e857-4c77-ad57-9e84c4356e4f
+begin
+	sta_b = @bind sta Select(stations, default=stations[1])
+	myurl="https://www.ndbc.noaa.gov/station_page.php?station=$(sta)"
+	md"""## Select Buoy
+	$(sta_b)"""
+end
+
+# ╔═╡ a1698e0e-db0d-4cd2-91b3-d530f77cd609
+begin
+    buoy=read(NOAAbuoy(),sta)
+	"Done with reading data"
+end
+
+# ╔═╡ c0a3ec23-2e76-40e6-bdf8-b6b774dabae5
+begin
+	nms=names(buoy.data)
+	nms=nms[findall([!ismissing(buoy.data[1,nm]) for nm in nms])]
+	nm0 = (in("PRES",nms) ?  "PRES" : nms[6])
+	var_b = @bind var Select(names(buoy.data), default=nm0)
+	md"""## Select variable 
+	$(var_b)"""
+end
+
[…]
