+++ v0.2.6/examples/Buoy_NWP_NOAA_monthly.jl
@@ -1,0 +1,1909 @@
+### A Pluto.jl notebook ###
+# v0.19.46
+
+using Markdown
+using InteractiveUtils
+
+# ╔═╡ 920b7a98-8478-4ad7-8897-8e5d92737199
+begin
+	using PlutoUI, CairoMakie, OceanRobots
+	"Done with packages"
+end
+
+# ╔═╡ 91db1379-9f47-4982-9716-f755451f819f
+begin
+img_url="https://nosc.noaa.gov/OSC/images/Kukui_and_NOAA_buoy_1.177308.JPG"
+	
+md"""# NOAA Buoys / Monthly Means"""
+end
+
+# ╔═╡ 2f87a4a9-ad5d-46ff-961b-028de3a27a23
+PlutoUI.TableOfContents()
+
+# ╔═╡ d106d4c1-1d2e-40b3-8958-901ea437a524
+begin
+md"""In this notebook, you will compute climatologies and distributions from NOAA buoy data.
+
+$(Resource(img_url,:width => 250))
+
+!!! note 
+	For more information about NOAA Buoys, visit <https://www.ndbc.noaa.gov/> and links below.
+
+## Visualize Data	
+"""
+end
+
+# ╔═╡ 502dde03-9ce3-4e15-bfc4-854c87a60026
+md"""## Data Summary
+
+Plot below shows:
+
+- top left: the full distribution of 25 year temperatue changes, using all individual months and years, is shown in the top left corner.
+- top right: averaged month by month, the 25 year change in temperature is positive for each month -- between +2 and +4 degrees, approximately.
+- bottom: the mean seasonal cycle in the ~1990s (blue) and ~2010s (orange). Range is ~ 25 degree in both cases. Shift between the curves is what's shown in top panel.
+"""
+
+# ╔═╡ 34a9a048-da3e-11ec-2017-b7b762895d86
+begin
+	ID=44013
+    buoy=read(NOAAbuoy_monthly(),ID)
+end
+
+# ╔═╡ 051ea6ce-a699-4a08-a4cf-214875bd3ca7
+plot(buoy;option=:demo)
+
+# ╔═╡ c821bb5e-ae74-4c06-8864-4202e7343e18
+md"""## Data Source
+
+- Station page : <https://www.ndbc.noaa.gov/station_page.php?station=44013>
+- Station history : <https://www.ndbc.noaa.gov/station_history.php?station=44013>
+- Thredds server : <https://dods.ndbc.noaa.gov/thredds/catalog/data/stdmet/44013/catalog.html>
+- Units : <https://www.ndbc.noaa.gov/measdes.shtml>
+- Climatic Summary Plots (SST, in degree C) : <https://www.ndbc.noaa.gov/view_climplot.php?station=44013&meas=st>
+
+## Appendix
+
+### Julia Packages
+"""
+
+# ╔═╡ e76df21f-e3ad-4c1e-97d6-3b775abd59ea
+md"""### Alternative Data Access Method
+
+!!! note
+    Using text files requires that you download them to a temporary folder as done below.
+"""
+
+# ╔═╡ aaf89c1f-bd6f-45d8-ae86-3c28bd0735ee
+let
[…]

[…]
