### 2024-09-24 #164 @luraess
> Can you point me where I can find "statement of need", "Community guidelines", "Results [...] compared to approaches with similar goals". @luraess are all these 3 additional points actually needed for the julia proceedings?

We are currently revising the guidelines for that and do no longer require explicit sections about it. We leave it up to the authors to state and motivate where appropriate the need for their work (and up to the reviewers to assess it).
