+++ docs/src/examples.md
@@ -25,7 +25,7 @@
 
 ## How-To
 
-To install `OceanRobots.jl` in `julia` proceed as usual via the package manager (`using Pkg; Pkg.add(OceanRobots)`).
+To install `OceanRobots.jl` in `julia` proceed as usual via the package manager (`using Pkg; Pkg.add("OceanRobots")`).
 
 To run a notebook interactively (`.jl` files) you want to use [Pluto.jl](https://github.com/fonsp/Pluto.jl). For example, copy and paste one of the above `code link`s in the [Pluto.jl interface](https://github.com/fonsp/Pluto.jl/wiki/🔎-Basic-Commands-in-Pluto). This will let you spin up the notebook in a web browser from the copied URL.
 
