### 2024-09-24 #164 @gaelforget
> Here are some additional comments/suggestions from my side:
>
> * I had some issues with the notebook “Global Ocean Metadata” which did not load for me at first...
> Maybe it would be good to tell the users to define `ENV["DATADEPS_ALWAYS_ACCEPT"] = "true"` or to download the file in a separate julia terminal.

Not sure why this works for me. Did you use the latest version of the notebook?

Since [v0.3.9 of MeshArrays.jl](https://github.com/JuliaClimate/MeshArrays.jl/commit/9a5d702cf69fbc09e574d932ca82a3972b716fa1) the download should happen automatically (see `withenv("DATADEPS_ALWAYS_ACCEPT"=>true) do` in MA_datadep). I just tried after manually deleting `datadeps/countries_shp1` and starting the notebook in Pluto -- the files did download automatically, so I dont know how to reproduce issue.

Just in case,  I re-updated the dependencies of the [notebook](https://github.com/JuliaOcean/OceanRobots.jl/blob/master/examples/OceanOPS.jl) (now using  `v0.3.16 of MeshArrays.jl` where `MA_datadep` is now `mydatadep`).

Could you please try again with the latest version and. confirm if issue persists?

> * Installation instructions at  “[https://juliaocean.github.io/OceanRobots.jl/dev/examples/”](https://juliaocean.github.io/OceanRobots.jl/dev/examples/%E2%80%9D). Replace
>   `using Pkg; Pkg.add(OceanRobots)`  by `using Pkg; Pkg.add(“OceanRobots”)`.

[done](https://github.com/JuliaOcean/OceanRobots.jl/commit/e78f37f93df339057a8221231a2a30fa05744091)

> * Functionality documentation: Can you point me where I can find the documentation of the public function?
>   There is a page for "internals" (https://juliaocean.github.io/OceanRobots.jl/dev/internals/#Internals) but I am not sure which of these functions is the public API.

Could be a matter of terminology but I think of the [docs API section](https://juliaocean.github.io/OceanRobots.jl/dev/reference/#User-Interface-(API)) as `documentation of the public function`.

Does this address your question or could you provide additional direction to do so?

The internals section may seem a bit disorganized as it includes all modules functions (`Modules = [OceanOPS, GDP, GDP_CloudDrift, NOAA, GliderFiles, ArgoFiles, OceanSites, CCHDO, SLA, podaac_sla, cmems_sla]`). On some level maybe the module names should be listed in the docs as `public function`.
