### 2024-09-25 #164 @Alexander-Barth
I can confirm that I have no-longer the datadeps issue!
Concerning the documentation, some doc strings only contain the function signature with an example:
https://github.com/JuliaOcean/OceanRobots.jl/blob/8b5f478ff5156dbbfedc0bfd52b9ab716e200445/ext/OceanRobotsMakieExt.jl#LL67-L75

"internal" versus "public" API: for me it is ok, not to document the internal API (or only minimally document them) as the user should only use  the public functions and the internal API is for developers and contributors. Thank you for pointing me to the section with the public functions.

I am wondering if you can expand the information.
Maybe you can explain briefly the type of plots that is produces (horizontal section with lon/lat axis, vertical section, vertical profile,...) and the positional and keyword additional arguments that are accepted. I guess for example that `size` is the width and height in pixels of the plot. Maybe that could be stated?

https://github.com/JuliaOcean/OceanRobots.jl/blob/8b5f478ff5156dbbfedc0bfd52b9ab716e200445/ext/OceanRobotsMakieExt.jl#LL129-L137

For keywords like `option=:demo` are other values of `option` possible? How do I know which values for `var` are valid? The doc string seem to indicated that per default `var = ""` but in the code there is `var="T(°F)"`. A lot becomes clear when you read the code, a typical use will normally see only the documentation.
