### 2024-09-29 d18e54e
use 41044 (no realtime data for 41046 currently)
