+++ Project.toml
@@ -7,7 +7,6 @@ version = "0.2.6"
 CFTime = "179af706-886a-5703-950a-314cd64e0468"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Dataverse = "9c0b9be8-e31e-490f-90fe-77697562404d"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
@@ -23,19 +22,18 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 URIs = "5c2747f8-b7ea-4ff2-ba2e-563bfd36b1d4"
 
 [weakdeps]
-Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 ArgoData = "9eb831cf-c491-48dc-bed4-6aca718df73c"
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 
 [extensions]
-OceanRobotsMakieExt = ["Makie"]
 OceanRobotsArgoDataExt = ["ArgoData"]
+OceanRobotsMakieExt = ["Makie"]
 
 [compat]
 ArgoData = "0.1"
 CFTime = "0.1"
 CSV = "0.10"
 DataFrames = "1"
-DataStructures = "0.18"
 Dataverse = "0.2"
 Downloads = "1"
 FTPClient = "1"
