### 2024-09-29 bb67b5d
rm old modules from docs list
