### 2024-09-29 1b536d8
fix previous commit, add NCDatasets to build docs
