+++ docs/Project.toml
@@ -5,6 +5,7 @@ DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Git = "d7ba0133-e1db-5d97-8f8c-041e4b3a1eb2"
 MeshArrays = "cb8c808f-1acf-59a3-9d2b-6e38d009f683"
+NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
 Pluto = "c3e4b0f8-55cb-11ea-2926-15256bba5781"
 PlutoSliderServer = "2fc8631c-6f24-4c5b-bca7-cbb509c42db4"
 Shapefile = "8e980c4a-a4fe-5da2-b3a7-4b4b0353a2f4"
