### 2024-09-30 cbe413b
add Climatology.jl to build docs
