+++ docs/src/reference.md
@@ -99,6 +99,6 @@ nothing #hide
 ```@example ex1
 using Climatology, CairoMakie, NCDatasets
 SLA=read(SeaLevelAnomaly(name="sla_podaac"))
-plot(sla)
+plot(SLA)
 ```
 
