+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"contributing/#Contributors-Guide","page":"Contribute","title":"Contributors Guide","text":"","category":"section"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"Thank you for considering contributing to OceanRobots.jl! If you're interested in contributing we want your help no matter how big or small a contribution you make! ","category":"page"},{"location":"contributing/#Table-of-Contents","page":"Contribute","title":"Table of Contents","text":"","category":"section"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"How to Contribute\nReporting Bugs\nSuggesting Enhancements\nCode Contribution\nPull Request Process\nLicense","category":"page"},{"location":"contributing/#How-to-Contribute","page":"Contribute","title":"How to Contribute","text":"","category":"section"},{"location":"contributing/#Reporting-Bugs","page":"Contribute","title":"Reporting Bugs","text":"","category":"section"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"If you encounter a bug, please help us fix it by following these steps:","category":"page"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"Ensure the bug is not already reported by checking the issue tracker.\nIf the bug isn't reported, open a new issue. Clearly describe the issue, including steps to reproduce it.","category":"page"},{"location":"contributing/#Suggesting-Enhancements","page":"Contribute","title":"Suggesting Enhancements","text":"","category":"section"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"If you have ideas for enhancements, new features, or improvements, we'd love to hear them! Follow these steps:","category":"page"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"Check the issue tracker to see if your suggestion has been discussed.\nIf not, open a new issue, providing a detailed description of your suggestion and the use case it addresses.","category":"page"},{"location":"contributing/#Code-Contribution","page":"Contribute","title":"Code Contribution","text":"","category":"section"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"If you'd like to contribute code to the project:","category":"page"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"Fork the repository.\nClone your fork: git clone https://github.com/juliaocean/OceanRobots.jl\nCreate a new branch for your changes: git checkout -b feature-branch\nMake your changes and commit them with a clear message.\nPush your changes to your fork: git push origin feature-branch\nOpen a pull request against the master branch of the main repository.","category":"page"},{"location":"contributing/#Pull-Request-Process","page":"Contribute","title":"Pull Request Process","text":"","category":"section"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"Please ensure your pull request follows these guidelines:","category":"page"},{"location":"contributing/","page":"Contribute","title":"Contribute","text":"Adheres to the coding standards.\nIncludes relevant tests for new functionality.\nHas a clear commit history and messages.\nReferences the re
[…]
