### 2024-10-01 faf6a44
improve docstrings to document parameters
