### 2024-10-01 #164 @gaelforget
> I can confirm that I have no-longer the datadeps issue! (Maybe I did not use the latest version of all dependencies)

Great to hear

> "internal" versus "public" API: for me it is ok, not to document the internal API (or only minimally document them) as the user should only use the public functions and the internal API is for developers and contributors. Thank you for pointing me to the section with the public functions.

ok. Thanks for the guideline. I commented out the internals part in https://github.com/JuliaOcean/OceanRobots.jl/pull/106

> Concerning the documentation, some doc strings only contain the function signature with an example: https://github.com/JuliaOcean/OceanRobots.jl/blob/8b5f478ff5156dbbfedc0bfd52b9ab716e200445/ext/OceanRobotsMakieExt.jl#LL67-L75
>
> I am wondering if you can expand the information. Maybe you can explain briefly the type of plots that is produces (horizontal section with lon/lat axis, vertical section, vertical profile,...) and the positional and keyword additional arguments that are accepted. I guess for example that `size` is the width and height in pixels of the plot. Maybe that could be stated?
>
> https://github.com/JuliaOcean/OceanRobots.jl/blob/8b5f478ff5156dbbfedc0bfd52b9ab716e200445/ext/OceanRobotsMakieExt.jl#LL129-L137
>
> For keywords like `option=:demo` are other values of `option` possible? How do I know which values for `var` are valid? The doc string seem to indicated that per default `var = ""` but in the code there is `var="T(°F)"`. A lot becomes clear when you read the code, a typical use will normally see only the documentation.

Thanks! The dosctrings had issues indeed. I updated them and added some detail. in https://github.com/JuliaOcean/OceanRobots.jl/pull/106

There is some variability in how the various data sets handle IDs for instrument/cruise/etc. I am realizing now that more needs to be said about this too in the docs.
