### 2024-10-07 51845bc
ArgoData (ext to dep, part 1)
