+++ src/files.jl
@@ -670,165 +670,6 @@ end
 
 ##
 
-module ArgoFiles
-
-using NCDatasets, Downloads, CSV, DataFrames, Interpolations, Statistics, Dates
-
-"""
-    ArgoFiles.list_floats(;list=DataFrame())
-
-Get list of Argo profilers from file `ArgoFiles.list_floats()`.
-
-Or write provided `list` to file as a DataFrame.
-
-```
-using OceanRobots, ArgoData
-ArgoFiles.list_floats(list=GDAC.files_list())
-```
-"""
-function list_floats(;list=DataFrame())
-    td=string(Dates.today())
-    fil=joinpath(tempdir(),"Argo_list_$(td).csv")
-    if isempty(list)
-        files_list=CSV.read(fil,DataFrame)
-    else
-        files_list=list
-        CSV.write(fil,files_list)
-        files_list
-    end
-end
-
-"""
-    ArgoFiles.download(files_list,wmo)
-
-Download an Argo profiler file.    
-"""
-function download(files_list,wmo)
-    ii=findall(files_list.wmo.==wmo)[1]
-    folder=files_list.folder[ii]
-
-    url0="https://data-argo.ifremer.fr/dac/$(folder)/"
-    fil=joinpath(tempdir(),"$(wmo)_prof.nc")
-
-    !isfile(fil) ? Downloads.download(url0*"/$(wmo)/$(wmo)_prof.nc",fil) : nothing
-
-    return fil
-end
-
-"""
-    ArgoFiles.readfile(fil)
-
-Read an Argo profiler file.    
-"""
-function readfile(fil)
-    ds=Dataset(fil)
-
-	lon=ds["LONGITUDE"][:]
-	lat=ds["LATITUDE"][:]
-
-	lon360=lon; lon[findall(lon.<0)].+=360
-	maximum(lon)-minimum(lon)>maximum(lon360)-minimum(lon360) ? lon=lon360 : nothing
-
-	PRES=ds["PRES_ADJUSTED"][:,:]
-	TEMP=ds["TEMP_ADJUSTED"][:,:]
-	PSAL=ds["PSAL_ADJUSTED"][:,:]
-	TIME=10*ones(size(PRES,1)).* (1:length(lon))' .-10.0
-    DATE=[DateTime(ds["JULD"][i]) for j in 1:size(PRES,1), i in 1:length(lon)]
-
-    close(ds)
-
-    return (lon=lon,lat=lat,PRES=PRES,TEMP=TEMP,PSAL=PSAL,TIME=TIME,DATE=DATE)
-end
-
-skmi(x) = ( sum((!ismissing).(x))>0 ? minimum(skipmissing(x)) : missing )
-skma(x) = ( sum((!ismissing).(x))>0 ? maximum(skipmissing(x)) : missing )
-
-"""
-    ArgoFiles.scan_txt(fil="ar_index_global_prof.txt"; do_write=false)
-
-Scan the Argo file lists and return summary tables in DataFrame format. 
-Write to csv file if `istrue(do_write)`.
-
-```
-ArgoFiles.scan_txt("ar_index_global_prof.txt",do_write=true)
-ArgoFiles.scan_txt("argo_synthetic-profile_index.txt",do_write=true)
-```
-"""
-function scan_txt(fil="ar_index_global_prof.txt"; do_write=false)
-    if fil=="ar_index_global_prof.txt"
-        filename=joinpath(tempdir(),"ar_index_global_prof.txt")
-        url="https://data-argo.ifremer.fr/ar_index_global_prof.txt"
-        outputfile=joinpath(tempdir(),"ar_index_global_prof.csv")
[…]
