+++ ext/OceanRobotsMakieExt.jl
@@ -180,88 +180,11 @@ function plot_summary(tbl,all,var,uni;size=(900,600))
 	f
 end
 
-## Argo
+##
 
-function heatmap_profiles!(ax,TIME,TEMP,cmap)
-	x=TIME[1,:]; y=collect(0.0:5:500.0)
-	co=Float64.(permutedims(TEMP))
-	rng=extrema(TEMP[:])
-	sca=heatmap!(ax, x , y , co, colorrange=rng,colormap=cmap)
-	ax.ylabel="depth (m)"
-	sca
-end
+#plot(x::ArgoFloat; kwargs...) = plot(ArgoData.OneArgoFloat(x.ID,x.data); kwargs...)
 
-function plot_profiles!(ax,TIME,PRES,TEMP,cmap)
-	ii=findall(((!ismissing).(PRES)).*((!ismissing).(TEMP)))
-
-	x=TIME[ii]
-	y=-PRES[ii] #pressure in decibars ~ depth in meters
-	co=Float64.(TEMP[ii])
-	rng=extrema(co)
-
-	sca=scatter!(ax, x , y ,color=co,colormap=cmap, markersize=5)
-
-	ax.xlabel="time (day)"
-	ax.ylabel="depth (m)"
-
-	sca
-end
-
-function plot_trajectory!(ax,lon,lat,co;
-		markersize=2,linewidth=3, pol=Any[],xlims=(-180,180),ylims=(-90,90),title="")
-	li=lines!(ax,lon, lat, linewidth=linewidth, color=co, colormap=:turbo)
-	scatter!(ax,lon, lat, marker=:circle, markersize=markersize, color=:black)
-	!isempty(pol) ? [lines!(ax,l1,color = :black, linewidth = 0.5) for l1 in pol] : nothing
-	ax.xlabel="longitude";  ax.ylabel="latitude"; ax.title=title
-	xlims!(xlims...); ylims!(ylims...)
-	li
-end
-
-xrng(lon)=begin
-	a=[floor(minimum(skipmissing(lon))) ceil(maximum(skipmissing(lon)))]
-	dx=max(diff(a[:])[1],10)
-	b=(a[2]>180 ? +180 : 0)
-	(max(a[1]-dx/2,-180+b),min(a[2]+dx/2,180+b))
-end
-yrng(lat)=begin
-	a=[floor(minimum(skipmissing(lat))) ceil(maximum(skipmissing(lat)))]
-	dx=max(diff(a[:])[1],10)
-	b=(max(a[1]-dx/2,-90),min(a[2]+dx/2,90))
-end
-
-function plot_standard(wmo,arr,spd,T_std,S_std; markersize=2,pol=Any[],size=(900,600))
-
-	xlims=xrng(arr.lon)
-	ylims=yrng(arr.lat)
-	
-	fig1=Figure(size=size)
-
-	ax=Axis(fig1[1,1])
-	li1=plot_trajectory!(ax,arr.lon,arr.lat,arr.TIME[1,:];
-		linewidth=5,pol=pol,xlims=xlims,ylims=ylims,
-		title="time since launch, in days")
-	Colorbar(fig1[1,2], li1, height=Relative(0.65))
-
-	ax=Axis(fig1[1,3])
-	li2=plot_trajectory!(ax,arr.lon,arr.lat,spd.speed;
-		linewidth=5,pol=pol,xlims=xlims,ylims=ylims,
-		title="estimated speed (m/s)")
-	Colorbar(fig1[1,4], li2, height=Relative(0.65))
-
-	ax=Axis(fig1[2,1:3],title="Temperature, °C")
-	hm1=heatmap_profiles!(ax,arr.DATE,T_std,:thermal)
-	Colorbar(fig1[2,4], hm1, height=Relative(0.65))
[…]
