### 2024-10-07 5939a66
ArgoData (ext to dep, part 2)
