+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["gaelforget <gforget@mit.edu>"]
 version = "0.2.7"
 
 [deps]
+ArgoData = "9eb831cf-c491-48dc-bed4-6aca718df73c"
 CFTime = "179af706-886a-5703-950a-314cd64e0468"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
@@ -22,11 +23,9 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 URIs = "5c2747f8-b7ea-4ff2-ba2e-563bfd36b1d4"
 
 [weakdeps]
-ArgoData = "9eb831cf-c491-48dc-bed4-6aca718df73c"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 
 [extensions]
-OceanRobotsArgoDataExt = ["ArgoData"]
 OceanRobotsMakieExt = ["Makie"]
 
 [compat]
