+++ docs/make.jl
@@ -1,8 +1,7 @@
-using Documenter, OceanRobots, PlutoSliderServer, CairoMakie, ArgoData
+using Documenter, OceanRobots, PlutoSliderServer, CairoMakie
 
 ENV["DATADEPS_ALWAYS_ACCEPT"]=true
 OceanRobotsMakieExt=Base.get_extension(OceanRobots, :OceanRobotsMakieExt)
-OceanRobotsArgoDataExt=Base.get_extension(OceanRobots, :OceanRobotsArgoDataExt)
 
 using MeshArrays, Shapefile, DataDeps
 pol_file=demo.download_polygons("ne_110m_admin_0_countries.shp")
@@ -11,7 +10,7 @@ pol=MeshArrays.read_polygons(pol_file)
 ##
 
 makedocs(;
-    modules=[OceanRobots, OceanRobotsMakieExt,OceanRobotsArgoDataExt],
+    modules=[OceanRobots, OceanRobotsMakieExt],
     format=Documenter.HTML(),
     pages=[
         "Home" => "index.md",
