+++ docs/make.jl
@@ -38,13 +38,6 @@ for i in lst
     cp(fil_in,fil_out[1:end-4]*"jl")
 end
 
-for fil in ["argo_synthetic-profile_index.txt", "ar_index_global_prof.txt"]
-    ArgoFiles.scan_txt(fil,do_write=true)
-    fil_in=joinpath(tempdir(),fil[1:end-4]*".csv")
-    fil_out=joinpath(@__DIR__,"build", fil[1:end-4]*".csv")
-    mv(fil_in,fil_out)
-end
-
 deploydocs(;
     repo="github.com/JuliaOcean/OceanRobots.jl",
 )
