### 2024-10-07 f6478d8
allow non numeric ID for NOAAbuoy, fix notebook (till v0.2.7 released)
