### 2024-11-08 #164 @gaelforget
> @jonathanlilly could you please react to this message and provide a time estimate for when you can hand in the revision? If you no longer have time, this is also fine, but please tell it - as we could then find another reviewer and not hold back further the process. I'll set a reminder in 4 days. Thanks

Thanks @luraess for the reminders. I was still hoping that @jonathanlilly could complete the checklist quickly, but I agree that we might need to find another reviewer to avoid even more delay.
