### 2024-11-08 #164 @gaelforget
> > @jonathanlilly could you please react to this message and provide a time estimate for when you can hand in the revision? If you no longer have time, this is also fine, but please tell it - as we could then find another reviewer and not hold back further the process. I'll set a reminder in 4 days. Thanks
>
> My apologies for the long delay, this had fallen off my radar. I'll try to get in done in the next few days.

That would be awesome!

And quite timely as I am schedule to give a presentation related to this work in a couple weeks @ https://www.oceanpredict24.org

Thank you in advance for your review @jonathanlilly
