+++ paper/ref.bib
@@ -1,4 +1,25 @@
 
+@article{Rousselet2021,
+author = {Louise Rousselet  and Paola Cessi  and Gael Forget},
+title = {Coupling of the mid-depth and abyssal components of the global overturning circulation according to a state estimate},
+journal = {Science Advances},
+volume = {7},
+number = {21},
+pages = {eabf5478},
+year = {2021},
+doi = {10.1126/sciadv.abf5478},
+}
+
+@article{Forget2019,
+author = {Forget, Ga{\"{e}}l and Ferreira, David},
+doi = {10.1038/s41561-019-0333-7},
+issn = {1752-0908},
+journal = {Nature Geoscience},
+title = {{Global ocean heat transport dominated by heat export from the tropical Pacific}},
+url = {https://doi.org/10.1038/s41561-019-0333-7},
+year = {2019}
+}
+
 @article{Bittig2018,
 AUTHOR={Bittig, Henry C. and Steinhoff, Tobias and Claustre, Hervé and Fiedler, Björn and Williams, Nancy L. and Sauzède, Raphaëlle and Körtzinger, Arne and Gattuso, Jean-Pierre},   
 TITLE={An Alternative to Static Climatologies: Robust Estimation of Open Ocean {CO2} Variables and Nutrient Concentrations From {T, S, and O2} Data Using Bayesian Neural Networks},      
@@ -10,7 +31,7 @@ DOI={10.3389/fmars.2018.00328},
 	ISSN={2296-7745},   
 }
 
-@article{Sauzède2017,
+@article{Sauzede2017,
   AUTHOR={Sauzède, Raphaëlle and Bittig, Henry C. and Claustre, Hervé and Pasqueron de Fommervault, Orens and Gattuso, Jean-Pierre and Legendre, Louis and Johnson, Kenneth S.},    
 TITLE={Estimates of Water-Column Nutrient Concentrations and Carbonate System Parameters in the Global Ocean: A Novel Approach Based on Neural Networks},      
 JOURNAL={Frontiers in Marine Science},      
