+++ paper/paper.tex
@@ -29,7 +29,7 @@ pdfkeywords = {Julia, Ocean, Robots, Observation, In Situ, Model},
 
 \begin{abstract}
 
-DTOR is a framework to access and simulate the global fleet of ocean observing devices (or {\it ocean robots}) that monitor climate change. It brings these observations to Julia and let's us pair ocean robots with virtual counterparts (or {\it twins}). Digital twins provide a bridge to predictive models that enables machine learning. In turn, simulating observations in a virtual environment can help evaluate observational strategies a priori, during deployment, or afterwards. In this paper we present the framework, review simulated ocean robots, and discuss envisioned applications. 
+DTOR is a framework to access and simulate the global fleet of ocean observing devices (or {\it ocean robots}) that monitor climate change. It brings these observations to Julia and let's us pair ocean robots with virtual counterparts (or {\it twins}). Digital twins provide a bridge to predictive models that enables machine learning. In turn, simulating observations in a digital environment can help evaluate observational strategies a priori, during deployment, or afterwards. In this paper we present the framework, review simulated ocean robots, and discuss envisioned applications. 
 
 %A fleet of ocean robots is collecting data that are crucial to monitor, understand, and predict climate change. Our {\it digital twins for ocean robots} framework brings these data to Julia and provides the ability to pair them with a virtual counterpart. The simulated ocean robots provide a bridge to numerical models and enable machine learning. In turn, the simulated observation of a virtual Earth can be used to evaluate observing system strategies ahead a priori, and once they are in operation. Our frameworks can readily simulates complex data sets being collected at sea. It leverages multiple Julia packages developed by the author and the community. The 2023 JuliaCon talk focused on the global observing system being simulated, as well as the expected scientific, societal, and commercial applications.
 
@@ -37,9 +37,9 @@ DTOR is a framework to access and simulate the global fleet of ocean observing d
 
 \section{Introduction}
 
-The DTOR framework was introduced at the Symposium on Advances in Ocean Observation in 2022, and later presented to US-CLIVAR and JuliaCon in 2023. The primary goal of DTOR is to associate ocean observing systems (or {\it ocean robots}) with virtual counterparts (or {\it twins}), and do it for the whole fleet of ocean robots that are currently at sea or have observed the Ocean in the past (Fig.~\ref{fig:ocean-ops}). The virtual twins are to be created through numerical model simulations, which can come in different flavors and languages. The simulation of observing systems in the future, as climate change progresses, is also part of the scope of the DTOR project. At JuliaCon 2023, we presented a solution to (1) access and manipulate the data collected by ocean robots (inc. \pkg{OceanRobots.jl} and \pkg{ArgoData.jl}), and (2) simulate such observations in a virtual environment using a hierarchy of models (incl. \pkg{ClimateModels.jl} and \pkg{MITgcm.jl}).
[…]
