### 2024-11-12 #164 @gaelforget
Regenerating pdf following package name update (Drifters.jl in place of IndividualDisplacements.jl),
plus another round of proofreading. This is the diff : https://github.com/JuliaOcean/OceanRobots.jl/compare/ddb52f8f144cccf986420b149ac33e5652ba3609...JuliaConProceedings2023
