### 2024-11-12 JuliaOcean/OceanRobots.jl#117 @jonathanlilly
error in running testset
I got the following error in running the testset

OceanRobots.jl: Error During Test at /Users/lilly/.julia/packages/OceanRobots/EJutr/test/runtests.jl:4
  Got exception outside of a @test
  UndefVarError: `ArgoFiles` not defined in `Main`
  Hint: It looks like two or more modules export different bindings with this name, resulting in ambiguity. Try explicitly importing it from a particular module, or qualifying the name with the module it should come from.
  Hint: a global variable of this name also exists in OceanRobots.
  Hint: a global variable of this name also exists in ArgoData.
  Stacktrace:
   [1] macro expansion
	 @ ~/.julia/packages/OceanRobots/EJutr/test/runtests.jl:38 [inlined]
   [2] macro expansion
	 @ ~/.julia/juliaup/julia-1.11.1+0.aarch64.apple.darwin14/share/julia/stdlib/v1.11/Test/src/Test.jl:1700 [inlined]
   [3] top-level scope
	 @ ~/.julia/packages/OceanRobots/EJutr/test/runtests.jl:6
   [4] include(fname::String)
	 @ Main ./sysimg.jl:38
   [5] top-level scope
	 @ none:6
   [6] eval
	 @ ./boot.jl:430 [inlined]
   [7] exec\_options(opts::Base.JLOptions)
	 @ Base ./client.jl:296
   [8] \_start()
	 @ Base ./client.jl:531
Test Summary:  | Pass  Error  Total     Time
OceanRobots.jl |    8      1      9  1m25.4s
ERROR: LoadError: Some tests did not pass: 8 passed, 0 failed, 1 errored, 0 broken.
in expression starting at /Users/lilly/.julia/packages/OceanRobots/EJutr/test/runtests.jl:4
ERROR: Package OceanRobots errored during testing

I also got a bunch of warnings like this

┌ Warning: thread = 1 warning: parsed expected 16 columns, but didn't reach end of line around data row: 37338. Parsing extra columns and widening final columnset
└ @ CSV ~/.julia/packages/CSV/XLcqT/src/file.jl:593
┌ Warning: thread = 1 warning: only found 16 / 22 columns around data row: 37339. Filling remaining columns with `missing`
└ @ CSV ~/.julia/packages/CSV/XLcqT/src/file.jl:592
┌ Warning: thread = 1 warning: only found 16 / 22 columns around data row: 37340. Filling remaining columns with `missing`
└ @ CSV ~/.julia/packages/CSV/XLcqT/src/file.jl:592
┌ Warning: thread = 1 warning: only found 16 / 22 columns around data row: 37341. Filling remaining columns with `missing`
└ @ CSV ~/.julia/packages/CSV/XLcqT/src/file.jl:592
