### 2024-11-12 JuliaOcean/OceanRobots.jl#118 @jonathanlilly
warning in running Code 1
I get a warning in running Code 1 from 10.21105.jcon.00164.pdf

┌ Warning: specifycing files_list here is deprecated
└ @ ArgoData ~/.julia/packages/ArgoData/s0ZAd/src/data_structures.jl:164
ArgoFloat(6900900, (lon = Union{Missing, Float64}[4.186, 4.141, 3.983, 3.623, 4.011, 5.111, 4.887, 4.643, 4.799, 5.324  …  120.285, 120.149, 120.589, 122.984, 124.966, 126.023, 126.654, 127.086, 127.082, 126.743], lat = Union{Missing, Float64}[-46.989, -46.902, -46.757, -47.048, -47.254, -46.836, -46.41, -46.262, -46.272, -46.309  …  -49.727, -48.739, -47.954, -48.316, -49.014, -47.845, -47.37, -47.213, -47.11, -47.063], PRES = Union{Missing, Float32}[9.0 6.0 … 6.0 6.0; 14.0 7.0 … 7.0 8.0; … ; missing missing … missing missing; missing missing … missing missing], TEMP = Union{Missing, Float32}[7.602 7.707 … 10.15 10.656; 7.601 7.708 … 10.15 10.617; … ; missing missing … missing missing; missing missing … missing missing], PSAL = Union{Missing, Float32}[33.848 33.851 … 34.569 34.558; 33.847 33.85 … 34.569 34.56; … ; missing missing … missing missing; missing missing … missing missing], TIME = [0.0 10.0 … 2520.0 2530.0; 0.0 10.0 … 2520.0 2530.0; … ; 0.0 10.0 … 2520.0 2530.0; 0.0 10.0 … 2520.0 2530.0], DATE = [DateTime("2015-02-13T09:29:00") DateTime("2015-02-15T05:15:00") … DateTime("2021-12-30T06:57:00") DateTime("2022-01-09T07:03:00"); DateTime("2015-02-13T09:29:00") DateTime("2015-02-15T05:15:00") … DateTime("2021-12-30T06:57:00") DateTime("2022-01-09T07:03:00"); … ; DateTime("2015-02-13T09:29:00") DateTime("2015-02-15T05:15:00") … DateTime("2021-12-30T06:57:00") DateTime("2022-01-09T07:03:00"); DateTime("2015-02-13T09:29:00") DateTime("2015-02-15T05:15:00") … DateTime("2021-12-30T06:57:00") DateTime("2022-01-09T07:03:00")]))
