### 2024-11-12 JuliaOcean/OceanRobots.jl#119 @jonathanlilly
error in running Code 2
I get an error in running Code 2.

F = FlowFields(u = u , v = v , period = T )
ERROR: UndefVarError: `u` not defined in `Main`
Suggestion: check for spelling errors or missing imports.
Stacktrace:
 [1] top-level scope
   @ REPL[15]:1
