### 2024-11-12 JuliaOcean/OceanRobots.jl#118 @jonathanlilly
Code 1 geneates a figure that looks somewhat different from that in Figure 4 of the paper.  In my version, the top two figures have a different colormap and also do not show the continents.

It would be good to have a default choice of a 1:1 aspect ratio at, say, the mean latitude.
