### 2024-11-12 JuliaOcean/OceanRobots.jl#120 @jonathanlilly
warning in opening notebooks
I had the following warning in opening the notebooks locally... I'm using the current release of Julia, 1.11.1

┌ Info: Operation failed. Updating registries and trying again...
│   exception =
│    Unsatisfiable requirements detected for package Statistics [10745b16]:
│     Statistics [10745b16] log:
│     ├─possible versions are: 1.11.0 - 1.11.1 or uninstalled
└     └─restricted to versions 1.10.0 by an explicit requirement — no versions left
    Updating registry at `~/.julia/registries/General.toml`
┌ Warning: Operation failed. Removing Manifest and trying again...
│   exception =
│    Unsatisfiable requirements detected for package Statistics [10745b16]:
│     Statistics [10745b16] log:
│     ├─possible versions are: 1.11.0 - 1.11.1 or uninstalled
│     └─restricted to versions 1.10.0 by an explicit requirement — no versions left
