### 2024-11-12 #164 @jonathanlilly
This is a useful package... particularly just having the tools to download all this data in the same place.

Some small comments on the paper

In the abstract, "let's" should be "lets".
The word choice "restful" on the 5th line of section 2 is odd.
sea level pressure along their trajectories.
On p 2, last line, time-variability should not be hyphenated
On p 4,  first column, there is a missing reference and partiucularly is misspelled.

Comments on the package

It would be good to have installation instructions in the Readme.  I think the information in the example HowTo would be better placed in the Readme.

I don't really think the term "Ocean Robots" is particularly apt.  This appears to include all major in situ observing platforms.  Of these, I would say drifters, buoys, and ships are definitely not robots.   I don't think the term is going to resonate with people, and consequently, I think that with the current name a lot of people who could use this notebook will not manage to find it.

Regarding Community guidelines: "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I see guidelines for 1) and 2) in /docs/src/contributing.md.  I don't see guidelines for seeking support.

I was able to verify the behavior of the notebooks locally.  However, I had a warning in Code 1 (https://github.com/JuliaOcean/OceanRobots.jl/issues/118#issue-2651655592), an error in Code 2 (https://github.com/JuliaOcean/OceanRobots.jl/issues/119#issue-2651689940), an error in running the testset (https://github.com/JuliaOcean/OceanRobots.jl/issues/117#issue-2651633397), and a warning when running the notebooks (https://github.com/JuliaOcean/OceanRobots.jl/issues/120#issue-2651829709).

@luraess, you may want to know that the format used for the paper makes it difficult to copy and paste the code examples.  When I do so I get text that looks like this, with extra spaces:

u s i n g O c e a n R o b o t s , A r g o D a t a , C a i r o M a k i e
l s t = G D A C . f i l e s _ l i s t ( ) ; w m o = 6 9 0 0 9 0 0
a r g o = r e a d ( A r g o F l o a t ( ) , w m o = w m o , f i l e s _ l i s t = l s t )
