### 2024-11-12 #164 @luraess
Thanks @jonathanlilly for your feedback. I will see if we can improve the font format I future versions in order to circumvent the copy-paste issue detailed here.

@gaelforget Going reading through the latest draft, I see you have a `.` instead of a space in the author name which you may want to address.
