+++ paper/paper.tex
@@ -29,7 +29,7 @@ pdfkeywords = {Julia, Ocean, Robots, Observation, In Situ, Model},
 
 \begin{abstract}
 
-DTOR is a framework to access and simulate the global fleet of ocean observing devices (or {\it ocean robots}) that monitor climate change. It brings these observations to Julia and let's us pair ocean robots with virtual counterparts (or {\it twins}). Digital twins provide a bridge to predictive models that enables machine learning. In turn, simulating observations in a digital environment can help evaluate observational strategies a priori, during deployment, or afterwards. In this paper we present the framework, review simulated ocean robots, and discuss envisioned applications. 
+DTOR is a framework to access and simulate the global fleet of ocean observing devices (or {\it ocean robots}) that monitor climate change. It brings these observations to Julia and lets us pair ocean robots with virtual counterparts (or {\it twins}). Digital twins provide a bridge to predictive models that enables machine learning. In turn, simulating observations in a digital environment can help evaluate observational strategies a priori, during deployment, or afterwards. In this paper we present the framework, review simulated ocean robots, and discuss envisioned applications. 
 
 %A fleet of ocean robots is collecting data that are crucial to monitor, understand, and predict climate change. Our {\it digital twins for ocean robots} framework brings these data to Julia and provides the ability to pair them with a virtual counterpart. The simulated ocean robots provide a bridge to numerical models and enable machine learning. In turn, the simulated observation of a virtual Earth can be used to evaluate observing system strategies ahead a priori, and once they are in operation. Our frameworks can readily simulates complex data sets being collected at sea. It leverages multiple Julia packages developed by the author and the community. The 2023 JuliaCon talk focused on the global observing system being simulated, as well as the expected scientific, societal, and commercial applications.
 
@@ -49,7 +49,7 @@ The model hierarchy includes fast climate model emulators, global model output, 
 
 \section{Robots Observing the Ocean}
 
-Part of the fleet of ocean robots currently at sea is depicted in Fig.~\ref{fig:ocean-ops}, focusing on some of the most common observing platforms. To create this map, \pkg{OceanRobots.jl} queries the meta-data-base from Ocean-OPS.org, which keeps track of the whole ocean fleet of scientific observation platforms in real time, and provides a restful API. The \pkg{OceanRobots.jl} package provides a simple interface to this API via the `OceanOPS` module. \pkg{OceanRobots.jl} defines data structures to access and utilize data from common ocean robot types such as  `SurfaceDrifter` and `ArgoFloat` (shown in Fig.~\ref{fig:ocean-ops}). A non-exhaustive list of ocean observing platform categories within the scope of \pkg{OceanRobots.jl} is provided in Tab.~\ref{tab:robots}.
[…]
