### 2024-11-14 #164 @gaelforget
> In the abstract, "let's" should be "lets".
The word choice "restful" on the 5th line of section 2 is odd.
sea level pressure along their trajectories.
On p 2, last line, time-variability should not be hyphenated
On p 4, first column, there is a missing reference and partiucularly is misspelled.

- [x]  In the abstract, "let's" should be "lets".
- [x] The word choice "restful" on the 5th line of section 2 is odd.
  - this word refers to https://restfulapi.net
  - have now capitalized the acronym part
- [x] sea level pressure along their trajectories.
  - replaced with `trajectory`, which I think was what you meant
- [x] On p 2, last line, time-variability should not be hyphenated
- [x] On p 4, first column, there is a missing reference and partiucularly is misspelled.
  -  fixed `partiucularly`
  -  could not find missing ref; if it's still there in the revised version below, please quote the sentence I should looking for
