### 2024-11-14 #164 @luraess
@gaelforget your author name still displays with a `.` which would be good to fix
