+++ paper/paper.tex
@@ -21,7 +21,7 @@
 \hypersetup{
 pdftitle = {Digital Twins for Ocean Robot},
 pdfsubject = {JuliaCon 2023 Proceedings},
-pdfauthor = {Gaël.Forget},
+pdfauthor = {Gaël Forget},
 pdfkeywords = {Julia, Ocean, Robots, Observation, In Situ, Model},
 }
 
