+++ paper/paper.tex
@@ -12,7 +12,7 @@
 
 \title{Digital Twins for Ocean Robots}
 
-\author[1]{Gaël.Forget}
+\author[1]{Gaël Forget}
 \affil[1]{Earth, Atmospheric and Planetary Sciences, Massachusetts Institute of Technology, \newline
 77 Massachusetts Avenue, Cambridge, 02139, MA, USA}
 
