### 2024-11-14 #164 @gaelforget
> @gaelforget your author name still displays with a `.` which would be good to fix
>
> <img alt="Screenshot 2024-11-14 at 11 36 56" width="435" src="https://private-user-images.githubusercontent.com/61313342/386157379-f63e822c-7100-484f-bdd8-c2fcb104bea8.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzE1OTA1MzcsIm5iZiI6MTczMTU5MDIzNywicGF0aCI6Ii82MTMxMzM0Mi8zODYxNTczNzktZjYzZTgyMmMtNzEwMC00ODRmLWJkZDgtYzJmY2IxMDRiZWE4LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDExMTQlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQxMTE0VDEzMTcxN1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWM3MzU5Y2U1ZjdjNTIxMTA2ZTRkN2Q2ZWZkODgyMDM1M2VmNmE1MWVmMzZjNzRmZmViOWZlYjNiMDkxNTA4MGUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.9wCp1ywETdJC56OG86e3m-P8yEy1-PgFlX5x55kKNQU">

Done. Thanks for pointing this out!
