### 2024-11-14 #164 @gaelforget
> Regarding Community guidelines: "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"
> I see guidelines for 1) and 2) in /docs/src/contributing.md. I don't see guidelines for seeking support.

Done, in 3681ddf1655635741b1901dd60f5c287342b89bc , in section that was renamed as `User Directions`
