### 2024-11-14 #164 @gaelforget
> I don't really think the term "Ocean Robots" is particularly apt. This appears to include all major in situ observing platforms. Of these, I would say drifters, buoys, and ships are definitely not robots. I don't think the term is going to resonate with people, and consequently, I think that with the current name a lot of people who could use this notebook will not manage to find it.

Not sure what do about this one.

- Maybe that depends on how you think of / define what robots are. "A machine or device that operates automatically or by remote control." is a definition that applies to drifters, buoys, or a CTD rosette in my view. Also I envision this package to include more AUVs and ROVs moving forward. So I like the name robots for these reasons.
- I feel there is tradeoff between a specialized audience, and a broader one. `Robots` speaks to everyone on some level, including high-school students and younger, even though it might not be the most technically accurate. I am open to suggestions but I don't have a better idea at this time for a name that would align better with my goals.
- Could rename package later but would prefer to keep the name as is, and in alignment with the Juliacon presentation.
