### 2024-11-14 #164 @gaelforget
> @luraess, you may want to know that the format used for the paper makes it difficult to copy and paste the code examples. When I do so I get text that looks like this, with extra spaces:

I noticed this too, would love to resolve this issue, but don't know how.

Is there a way to avoid the blank spaces @luraess ? is it that I messed up on my end somehow?
