### 2024-11-14 #164 @gaelforget
I was able to verify the behavior of the notebooks locally. However, I had a

- [ ] warning in Code 1 (https://github.com/JuliaOcean/OceanRobots.jl/issues/118#issue-2651655592)
- [ ]  an error in Code 2 (https://github.com/JuliaOcean/OceanRobots.jl/issues/119#issue-2651689940)
- [ ] an error in running the testset (https://github.com/JuliaOcean/OceanRobots.jl/issues/117#issue-2651633397)
- [ ]  a warning when running the notebooks (https://github.com/JuliaOcean/OceanRobots.jl/issues/120#issue-2651829709).
