### 2024-11-14 #164 @luraess
> Is there a way to avoid the blank spaces @luraess ? is it that I messed up on my end somehow?

Could be an issue with the font used in the template. I could check that and let you know.
