### 2024-11-15 #164 @luraess
I just checked and it is not an issue on your side. Realistically, we won't have bandwidth to change something soon enough wrt this getting published so I would just keep it as such for now. It is anyway challenging to select code from snippet via copy & paste because of the two-column layout. I will discuss with the others EiCs to see how we can improve the template in the future.
