### 2024-11-17 a2c3d78
update example to avoid warning
