### 2024-11-17 d2970c8
get SurfaceDrifter from http server (instead of failing ftp server)
