### 2024-11-17 7ab775f
docs improvement from JCON proceeding paper review
