+++ docs/src/examples.md
@@ -21,34 +21,3 @@
 
 - For Argo and state estimates, see [ArgoData.jl](https://github.com/JuliaOcean/ArgoData.jl)
 - For simulations of drifter data, see [IndividualDisplacements.jl](https://github.com/JuliaClimate/IndividualDisplacements.jl)
-
-## How-To
-
-To install `OceanRobots.jl` in `julia` proceed as usual via the package manager (`using Pkg; Pkg.add("OceanRobots")`).
-
-To run a notebook interactively (`.jl` files) you want to use [Pluto.jl](https://github.com/fonsp/Pluto.jl). For example, copy and paste one of the above `code link`s in the [Pluto.jl interface](https://github.com/fonsp/Pluto.jl/wiki/🔎-Basic-Commands-in-Pluto). This will let you spin up the notebook in a web browser from the copied URL.
-
-All you need to do beforehand is to install [julia](https://julialang.org) and `Pluto.jl`. The installation of OceanRobots.jl and other Julia packages will then happen automatically when you run the notebook. 
-
-You can also download the notebooks folder and run them as normal Julia programs. We recommend runing each notebook in its own environment as shown below. 
-
-!!! note
-    To download OceanRobots.jl folder, which includes the notebooks folder, you can use `Git.jl`.
-
-```
-using Pkg; Pkg.add("Git"); using Git
-url="https://github.com/JuliaOcean/OceanRobots.jl"
-run(`$(git()) clone $(url)`)
-```
-
-```@example 1
-using Pkg; Pkg.add("Pluto"); using Pluto
-
-notebook="MITgcm.jl/examples/Float_Argo.jl"
-import OceanRobots; path=dirname(dirname(pathof(OceanRobots))) #hide
-notebook=joinpath(path,"examples","Float_Argo.jl") #hide
-Pluto.activate_notebook_environment(notebook)
-Pkg.instantiate()
-include(notebook)
-Pkg.activate("..") #hide
-```
