### 2024-11-18 JuliaOcean/OceanRobots.jl#118 @gaelforget
- Warning has been resolved so I will close this issue
- the 1:1 ratio seems like a good idea but not sure how do to it and is more of a feature request (dont hesitate to send a PR or open a new issue to discuss specifics)
- on the colormap & polygons on the plots : I tend to change my mood about these graphical details, and may change again soon, so I dont really see this as an issue v the paper
