### 2024-11-18 JuliaOcean/OceanRobots.jl#119 @gaelforget
Good point!

Unfortunately, I am going to have to get back to this one when back from this week's conference though, since I left the `u,v` file and script back home on a different computer
