### 2024-11-18 JuliaOcean/OceanRobots.jl#117 @gaelforget
I think this has to do a ftp server being down. Will try again tomorrow ...
