### 2024-11-18 #164 @gaelforget
> an error in Code 2 ([JuliaOcean/OceanRobots.jl#119 (comment)](https://github.com/JuliaOcean/OceanRobots.jl/issues/119#issue-2651689940))

As said in https://github.com/JuliaOcean/OceanRobots.jl/issues/119 , but just to explain slight delay on my side.

Unfortunately, I am going to have to get back to this one when back from this week's conference though, since I left the u,v file and script back home on a different computer
