+++ paper/paper.tex
@@ -68,13 +68,11 @@ surface drifters & \struct{SurfaceDrifter}, \struct{CloudDrift} \\
 drifting profilers & \struct{ArgoFloat} \\
 moored buoys & \struct{OceanSite}, \struct{NOAAbuoy} \\
 sea gliders & \struct{Gliders} \\
-expendable bathythermographs & \tbd \\
-sea gliders & \struct{Gliders} \\
+expendable bathy-thermographs & \tbd \\
 sail drones & \tbd \\
-research vessels & CCHDO \\
-ships of opportunity & \tbd \\
-orbiting satellites & \struct{SeaLevelAnomaly} \\
-suborbital flights & \tbd \\
+\hline
+research vessel data & CCHDO \\
+ships of opportunity data & \tbd \\
 marine mammals & \tbd \\\hline
 \end{tabular}}
 \label{tab:robots}
