+++ paper/paper.tex
@@ -71,7 +71,7 @@ sea gliders & \struct{Gliders} \\
 expendable bathy-thermographs & \tbd \\
 sail drones & \tbd \\
 \hline
-research vessel data & CCHDO \\
+research vessel data & \struct{CCHDO} \\
 ships of opportunity data & \tbd \\
 marine mammals & \tbd \\\hline
 \end{tabular}}
