### 2024-11-24 4b3193a
update notebook dependencies
