### 2024-11-24 69c5452
docs on polygons
