+++ v0.2.8/examples/OceanOPS.jl
@@ -1,0 +1,2213 @@
+### A Pluto.jl notebook ###
+# v0.20.3
+
+using Markdown
+using InteractiveUtils
+
+# This Pluto notebook uses @bind for interactivity. When running this notebook outside of Pluto, the following 'mock version' of @bind gives bound variables a default value (instead of an error).
+macro bind(def, element)
+    #! format: off
+    quote
+        local iv = try Base.loaded_modules[Base.PkgId(Base.UUID("6e696c72-6542-2067-7265-42206c756150"), "AbstractPlutoDingetjes")].Bonds.initial_value catch; b -> missing; end
+        local el = $(esc(element))
+        global $(esc(def)) = Core.applicable(Base.get, el) ? Base.get(el) : iv(el)
+        el
+    end
+    #! format: on
+end
+
+# ╔═╡ ccf98691-9386-41b9-a957-3cdeba51312b
+using OceanRobots, CairoMakie, PlutoUI, MeshArrays
+
+# ╔═╡ 30277358-0a8d-437e-9d2e-ebc7c307db31
+using Shapefile, GeoJSON, DataDeps, PrettyTables, Proj
+
+# ╔═╡ 5fa93c17-0a01-44c6-8679-d712c786907a
+md"""# OceanOPS : Global Ocean Metadata
+
+Source : <https://www.ocean-ops.org/board>, <https://www.ocean-ops.org/share/>
+
+## Visualize Data Cover
+"""
+
+# ╔═╡ 31fe91d7-dea2-49e3-902b-31aa62075986
+TableOfContents()
+
+# ╔═╡ 1042f70c-4337-4bf2-b533-edf27a422365
+md"""## Select a Platform Type
+
+Each color represents one type of observing platform.
+"""
+
+# ╔═╡ 596bce95-e13f-4439-858f-e944834c0924
+begin
+	bind_nam = @bind nam Select([:Argo,:Drifter])
+	md"""## Explore Platform Metadata
+
+	Here you can select a data set and then a platform. Meta-data from that platform is displayed as a result.
+	
+	Data set : $(bind_nam)
+	"""
+end
+
+# ╔═╡ aa80092c-80b9-489c-97b9-06c3d39ac594
+begin
+	list_data=OceanOPS.get_list(nam)
+	bind_id = @bind id Select(list_data)
+	md"""Platform ID : $(bind_id)"""
+end
+
+# ╔═╡ 401180a9-cb62-4dc6-b0a1-35df35f834db
+begin
+	meta=OceanOPS.get_platform(id)
+	md"""
+	| Item         | Value |
+	|--------------|:-----------|
+	| platform ID  | $(meta.id) |
+	| status      | $(meta.status) |
+	| country      | $(meta.country) |
+	| ship      | $(meta.ship) |
+	| deployed      | $(meta.deployed) |
+	"""
+end
+
+# ╔═╡ cb5cce8c-f67f-496b-ba10-38eefe9285e1
+md"""## List of Platform Types"""
+
+# ╔═╡ 9d4b4eed-2cc8-49cd-bb4d-2e7d51e9e4d9
+begin
+	list_platform_types=OceanOPS.list_platform_types()
+
+	tb=pretty_table(
[…]
