### 2024-11-27 a1032ab
code that prepares Drifters_example.jld2 for JCON paper
