### 2024-11-27 648001a
update code snippets
