### 2024-11-28 JuliaOcean/OceanRobots.jl#117 @gaelforget
has been resolved ( see CI -- https://github.com/JuliaOcean/OceanRobots.jl/actions/workflows/ci.yml )
