### 2024-11-28 JuliaOcean/OceanRobots.jl#119 @gaelforget
Has been resolved by the latest versions of `Drifters` (v0.6.2) and `OceanRobots` (v0.2.8).

- The code below is now executed as [part of the documentation](https://juliaclimate.github.io/Drifters.jl/dev/workflow/) of `Drifters.jl` which has successfully build on GitHub.
- The `Drifters.Gulf_of_Mexico_setup` function downloads a [file with flow fields](https://doi.org/10.5281/zenodo.14229203) described in the paper.
- The docs of Drifters.jl also contain the plotting code.

```
using JLD2, CairoMakie, Drifters, OceanRobots;
P=Drifters.Gulf_of_Mexico_setup();
F=FlowFields(u=P.u,v=P.v,period=P.T);
I=Individuals(F,P.x0,P.y0);
[solve!(I,P.T .+P.dT*(n-1)) for n in 1:P.nt]
```

The same code should work on any local computer with a connection (since it works on GitHub).

Can you please confirm this works for you and/or close the issue @jonathanlilly ?
