### 2024-11-28 JuliaOcean/OceanRobots.jl#120 @gaelforget
Thanks for reporting this but that's normal behavior I would say.

Since I have [updated the notebook environments on Julia 1.11](https://github.com/JuliaOcean/OceanRobots.jl/commit/4b3193afa785576551883360a76e76c4d1540d54) you should no longer see the warning. However, users of Julia 1.10 would now get this warning I think.

There can only be one environment in each notebook. But this is generally benign, so that's ok.
