### 2024-11-28 #164 @gaelforget
I feel we may be all set after confirming that the two following items have been adressed satisfactorily :

- @Alexander-Barth  : `Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?`
- @jonathanlilly : https://github.com/JuliaOcean/OceanRobots.jl/issues/119#issue-2651689940

Please let me know if anything else remains outstanding.

Thanks!
