+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"contributing/#User-Directions","page":"User Directions","title":"User Directions","text":"","category":"section"},{"location":"contributing/#Table-of-Contents","page":"User Directions","title":"Table of Contents","text":"","category":"section"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"How to Install and Use\nHow to Seek Support\nHow to Contribute\nReporting Bugs\nSuggesting Enhancements\nCode Contribution\nPull Request Process\nLicense","category":"page"},{"location":"contributing/#how-to-install-use","page":"User Directions","title":"How to Install and Use","text":"","category":"section"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"To install OceanRobots.jl in julia proceed as usual via the package manager (using Pkg; Pkg.add(\"OceanRobots\")).","category":"page"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"To run a notebook interactively (.jl files) you want to use Pluto.jl. For example, copy and paste one of the above code links in the Pluto.jl interface. This will let you spin up the notebook in a web browser from the copied URL.","category":"page"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"All you need to do beforehand is to install julia and Pluto.jl. The installation of OceanRobots.jl and other Julia packages will then happen automatically when you run the notebook. ","category":"page"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"You can also download the notebooks folder and run them as normal Julia programs. We recommend runing each notebook in its own environment as shown below. ","category":"page"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"note: Note\nTo download OceanRobots.jl folder, which includes the notebooks folder, you can use Git.jl.","category":"page"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"using Pkg; Pkg.add(\"Git\"); using Git\nurl=\"https://github.com/JuliaOcean/OceanRobots.jl\"\nrun(`$(git()) clone $(url)`)","category":"page"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"using Pkg; Pkg.add(\"Pluto\"); using Pluto\n\nnotebook=\"MITgcm.jl/examples/Float_Argo.jl\"\nimport OceanRobots; path=dirname(dirname(pathof(OceanRobots))) #hide\nnotebook=joinpath(path,\"examples\",\"Float_Argo.jl\") #hide\nPluto.activate_notebook_environment(notebook)\nPkg.instantiate()\ninclude(notebook)\nPkg.activate(\"..\") #hide","category":"page"},{"location":"contributing/#how-to-seek-support","page":"User Directions","title":"How to Seek Support","text":"","category":"section"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"If something is unclear or proves difficult to use, please seek support by opening an issue on the repository.","category":"page"},{"location":"contributing/#how-to-contribute","page":"User Directions","title":"How to Contribute","text":"","category":"section"},{"location":"contributing/","page":"User Directions","title":"User Directions","text":"Thank you for considering contributing to OceanRobots.jl! If you're interested in contributing we want your help no matter how big or small a contribution you make! ","category":"page"},{"locatio
[…]
