### 2024-11-28 #164 @Alexander-Barth
Yes, the documentation is much more detailed now. Very nice!
