### 2024-11-28 #164 @luraess
Thanks for the latest inputs. Given that both reviewers have their checklist completed, I would suggest we could proceed. I'll generate the acceptance dry run and will let you some time to make a final check of the generated material. We could, after that, finalise the procedure.
