### 2024-11-28 #164 @luraess
@gaelforget are the current v0.2.8 and corresponding Zenodo archive the latest?
