### 2024-11-29 53a560b
fix case when WHP netCDF is not included
