### 2024-11-29 7f7fc68
add missing DOIs
