+++ paper/ref.bib
@@ -173,7 +173,8 @@ year = {2015}
   volume = {submitted},
   author = {Ga{\"{e}}l Forget},
   title = {{MITgcm.jl: a Julia interface to the MITgcm}},
-  journal = {Journal of Open Source Software}
+  journal = {Journal of Open Source Software},
+  doi = {10.21105/joss.06710},
 }
 
 
@@ -181,7 +182,8 @@ year = {2015}
  author = {Forget, Ga{\"{e}}l},
  title = {Energy Imbalance in the Sunlit Ocean Layer},
  number = {submitted},
- year = {2024}
+ year = {2024},
+ doi = {10.21203/rs.3.rs-3979671/v1},
 }
 
 @article{Forget2015a,
@@ -251,7 +253,8 @@ number = {C3},
 pages = {5753--5766},
 title = {{A finite-volume, incompressible Navier Stokes model for studies of the ocean on parallel computers}},
 volume = {102},
-year = {1997}
+year = {1997},
+ doi = {10.1029/96jc02775},
 }
 
 @article{camp-etal:04,
@@ -260,7 +263,8 @@ journal = {Ocean Modelling},
 pages = {221--244},
 title = {{Conservation of properties in a free surface model}},
 volume = {6},
-year = {2004}
+year = {2004},
+doi = {10.1016/s1463-5003(03)00009-x},
 }
 
 @article{HHG05,
@@ -271,7 +275,8 @@ pages = {1356--1371},
 publisher = {Elsevier},
 title = {{An efficient exact adjoint of the parallel MIT general circulation model, generated via automatic differentiation}},
 volume = {21},
-year = {2005}
+year = {2005},
+ doi = {10.1016/j.future.2004.11.010},
 }
 
 
