### 2024-11-29 #164 @gaelforget
> @gaelforget are the current v0.2.8 and corresponding Zenodo archive the latest?

Let's use Version `v0.2.9` instead if that's ok. I released it last night and improved the meta data a bit. Here is the bib entry :

```
@software{gael_forget_2024_14242208,
  author       = {Gael Forget},
  title        = {{OceanRobots.jl (v0.2.9) : a Julia package to
                   access, analyze, and simulate data generated by
                   ocean robots such as Argo floats and drifting
                   buoys}},
  month        = nov,
  year         = 2024,
  publisher    = {Zenodo},
  version      = {v0.2.9},
  doi          = {10.5281/zenodo.14242208},
  url          = {https://doi.org/10.5281/zenodo.14242208}
}
```
