### 2024-11-29 d1ea2df
final proofreading and edits (part 1)
