+++ paper/paper.tex
@@ -17,7 +17,7 @@
 \affil[1]{Earth, Atmospheric and Planetary Sciences, Massachusetts Institute of Technology, \newline
 77 Massachusetts Avenue, Cambridge, 02139, MA, USA}
 
-\keywords{Julia, Ocean, Climate, Robot, Observation, Platform, Sensor, Drifter, Buoy, Profiler, Satellite, Modeling, Artificial Intelligence, Data Assimilation, Parameter Inference, Climatology, Geospatial, Statistics}
+\keywords{Julia, Ocean, Climate, Robot, Observation, Observing, Platform, Sensor, Drifter, Buoy, Profiler, Model, Modeling, Machine Learning, Artificial Intelligence, Data Assimilation, Parameter Inference, Climatology, Geospatial, Statistics}
 
 \hypersetup{
 pdftitle = {Digital Twins for Ocean Robot},
@@ -30,7 +30,7 @@ pdfkeywords = {Julia, Ocean, Robots, Observation, In Situ, Model},
 
 \begin{abstract}
 
-DTOR is a framework to access and simulate the global fleet of ocean observing devices (or {\it ocean robots}) that monitor climate change. It brings these observations to Julia and lets us pair ocean robots with virtual counterparts (or {\it twins}). Digital twins provide a bridge to predictive models that enables machine learning. In turn, simulating observations in a digital environment can help evaluate observational strategies a priori, during deployment, or afterwards. In this paper we present the framework, review simulated ocean robots, and discuss envisioned applications. 
+DTOR is a framework to access, analyze, and simulate the global fleet of ocean observing devices (or {\it ocean robots}) that monitor climate change. It brings these observations to Julia and lets us pair ocean robots with virtual counterparts (or {\it twins}). Digital twins provide a bridge from the data to predictive models and enable machine learning. In turn, observing system simulations in a digital environment can help evaluate observational strategies a priori, during deployment, or afterwards. In this paper we present the DTOR framework, its supported observing systems, capabilities to simulate ocean robots, and envisioned applications. 
 
 %A fleet of ocean robots is collecting data that are crucial to monitor, understand, and predict climate change. Our {\it digital twins for ocean robots} framework brings these data to Julia and provides the ability to pair them with a virtual counterpart. The simulated ocean robots provide a bridge to numerical models and enable machine learning. In turn, the simulated observation of a virtual Earth can be used to evaluate observing system strategies ahead a priori, and once they are in operation. Our frameworks can readily simulates complex data sets being collected at sea. It leverages multiple Julia packages developed by the author and the community. The 2023 JuliaCon talk focused on the global observing system being simulated, as well as the expected scientific, societal, and commercial applications.
 
@@ -38,9 +38,9 @@ DTOR is a framework to access and simulate the global fleet of ocean observing d
 
 \section{Introduction}
 
[…]
