### 2024-11-29 745e317
final proofreading and edits (part 2)
