+++ paper/paper.tex
@@ -161,9 +161,9 @@ A digital twin is a set of virtual information constructs that mimics the struct
 
 In the {\it digital twins for ocean robots} framework (DTOR), interactivity is facilitated by Julia and its large ecosystem of software packages. In particular, \pkg{Makie.jl} \cite{Makie} and \pkg{Pluto.jl} \cite{Pluto} let us operate even complex modeling workflows from notebooks and apps. The tutorial examples in \pkg{OceanRobots.jl}, \pkg{ArgoData.jl}, \pkg{ClimateModels.jl}, and \pkg{MITgcm.jl} \cite{Forget2024b} are notably available as Pluto notebooks. Interaction between model and data is further enabled by the rich Julia ecosystem for machine learning, data assimilation, parameter inference, etc. Fig.~\ref{fig:MLP} provides an example where \pkg{Flux.jl} \cite{Innes2018} is used to train neural networks to predict chlorophyll concentration (linked to marine microbe abundance) from environmental parameters as done in the CANYON model \cite{Sauzede2017,Bittig2018}.
 
-The core of DTOR is formed by \pkg{OceanRobots.jl}, \pkg{ArgoData.jl}, and \pkg{MeshArrays.jl} for the physical twins, along with \pkg{Climatology.jl}, \pkg{ClimateModels.jl}, \pkg{MITgcm.jl}, and \pkg{Drifters.jl} for the virtual twins (i.e., predictive modeling). The \pkg{ClimateModels.jl} interface streamlines the use of models implemented in various languages. The model hierarchy already includes the models used in this paper, and is easy to extend via pkg{ClimateModels.jl}. DTOR can notably simulate observations in the future based on climate model predictions, and \pkg{ClimateModels.jl} provides two options for this kind of applications -- either querying the CMIP6 archive of model output \cite{CMIP6} or using a fast emulator such as Hector \cite{Hector2015} (Fig.~\ref{fig:globalwarming}). 
+The core of DTOR is formed by \pkg{OceanRobots.jl}, \pkg{ArgoData.jl}, and \pkg{MeshArrays.jl} for the physical twins, along with \pkg{Climatology.jl}, \pkg{ClimateModels.jl}, \pkg{MITgcm.jl}, and \pkg{Drifters.jl} for the virtual twins (i.e., predictive modeling). The \pkg{ClimateModels.jl} interface streamlines the use of models implemented in various languages. The model hierarchy already includes models listed in this paper, and is easy to extend via the \pkg{ClimateModels.jl} interface. DTOR can also simulate observations in the future based on climate model predictions (Fig.~\ref{fig:globalwarming}), and \pkg{ClimateModels.jl} provides two options for this kind of applications -- either querying the CMIP6 archive of model output \cite{CMIP6} or using a fast emulator such as Hector \cite{Hector2015}. 
 
[…]
