### 2024-11-29 5b43eb4
final proofreading and edits (part 3)
