+++ paper/paper.tex
@@ -40,7 +40,7 @@ DTOR is a framework to access, analyze, and simulate the global fleet of ocean o
 
 The DTOR framework was introduced at the Symposium on Advances in Ocean Observation in 2022, and later presented to US-CLIVAR and JuliaCon in 2023, and the OceanPredict Symposium in 2024. The primary goal of DTOR is to associate ocean observing systems (or {\it ocean robots}) with virtual counterparts (or {\it twins}), and do it for the whole fleet of ocean robots that are currently at sea or have observed the Ocean in the past (Fig.~\ref{fig:ocean-ops}). The virtual twins are to be created through numerical model simulations, which can come in different flavors and languages. The simulation of observing systems in the future, as climate change progresses, is also part of the scope of the DTOR project. At JuliaCon 2023, we presented a solution to (1) access and manipulate the data collected by ocean robots (inc. \pkg{OceanRobots.jl} and \pkg{ArgoData.jl}), and (2) simulate such observations in a digital environment using a hierarchy of models (incl. \pkg{ClimateModels.jl} and \pkg{MITgcm.jl}).
 
-The model hierarchy includes fast climate model emulators, global model output, ocean reanalyses, high-resolution models, and several ways to represent marine ecosystems. Through a streamlined workflow, \pkg{ClimateModels.jl} (JuliaCon 2021, 2023) makes it easy to operate these models that can provide a digital environment for the virtual ocean robot fleet to observe and navigate. \pkg{Drifters.jl} \cite{Forget2021} can be use to predict pathways of ocean robots that tend to follow ocean currents. \pkg{MeshArrays.jl} adds basic geospatial support for global climate model grids, providing capabilities such as interpolation and geolocation on a grid. The framework leverages and links to a variety of highly capable Julia packages from the community. It notably provides user-friendly plotting recipes via its \pkg{Makie.jl} extension, and tutorial examples in the form of \pkg{Pluto.jl} notebooks.
+The model hierarchy includes fast climate model emulators, global model output, ocean reanalyses, high-resolution models, and several ways to represent marine ecosystems. Through a streamlined workflow, \pkg{ClimateModels.jl} (JuliaCon 2021, 2023) makes it easy to operate these models that can provide a digital environment for the virtual ocean robot fleet to observe and navigate. \pkg{Drifters.jl} \cite{Forget2021} can be used to predict pathways of ocean robots that tend to follow ocean currents. \pkg{MeshArrays.jl} adds basic geospatial support for global climate model grids, providing capabilities such as interpolation and geolocation on a grid. The framework leverages and links to a variety of highly capable Julia packages from the community. It notably provides user-friendly plotting recipes via its \pkg{Makie.jl} extension, and tutorial examples in the form of \pkg{Pluto.jl} notebooks.
 
 \begin{figure}[t]
 \centerline{\includegraphics[width=\columnwidth]{figs/20240528_OceanOPS.png}}
@@ -58,7 +58,7 @@ Part of the fleet of ocean robots currently at sea is depicted in Fig.~\ref{fig:
 \label{fig:drifters}
 \end{figure}
 
[…]
