### 2024-11-29 #164 @gaelforget
> 👋 @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/JuliaCon/proceedings-papers/jcon.00164/jcon.00164/10.21105.jcon.00164.pdf)
>
> If the paper PDF and the deposit XML files look good in [JuliaCon/proceedings-papers#109](https://github.com/JuliaCon/proceedings-papers/pull/109), then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`

Done with final checks / proofreading -- [here is the diff](https://github.com/JuliaOcean/OceanRobots.jl/compare/b41911a0de07b55f2833f85e351ff5ddfc83e1b2..JuliaConProceedings2023)
