+++ paper/paper.tex
@@ -43,8 +43,8 @@ The DTOR framework was introduced at the Symposium on Advances in Ocean Observat
 The model hierarchy includes fast climate model emulators, global model output, ocean reanalyses, high-resolution models, and several ways to represent marine ecosystems. Through a streamlined workflow, \pkg{ClimateModels.jl} (JuliaCon 2021, 2023) makes it easy to operate these models that can provide a digital environment for the virtual ocean robot fleet to observe and navigate. \pkg{Drifters.jl} \cite{Forget2021} can be used to predict pathways of ocean robots that tend to follow ocean currents. \pkg{MeshArrays.jl} adds basic geospatial support for global climate model grids, providing capabilities such as interpolation and geolocation on a grid. The framework leverages and links to a variety of highly capable Julia packages from the community. It notably provides user-friendly plotting recipes via its \pkg{Makie.jl} extension, and tutorial examples in the form of \pkg{Pluto.jl} notebooks.
 
 \begin{figure}[t]
-\centerline{\includegraphics[width=\columnwidth]{figs/20240528_OceanOPS.png}}
-\caption{Location of ocean robots that were collecting data on {\it 2024/05/27} for three types of observations (blue, green, and red dots). Deployments of Argo profilers that were being planned at the time are also shown (purple dots). Data is from https://www.ocean-ops.org and queried via \pkg{OceanRobots.jl}.}
+\centerline{\includegraphics[width=\columnwidth]{figs/20241129_OceanOPS.png}}
+\caption{Ocean robots that were collecting data on {\it 2024/11/29} for three types of observations (blue, green, and red dots). Planned deployments of Argo profilers are also shown (purple dots). Data is from https://www.ocean-ops.org and queried via \pkg{OceanRobots.jl}.}
 \label{fig:ocean-ops}
 \end{figure}
 
