+++ examples/OceanOPS.jl
@@ -130,35 +130,35 @@ begin
 end
 
 # ╔═╡ c5f24071-1c1f-4edf-b7b2-b6194c33b571
-let
+begin
 	function demofigure()
 		lon0=-160
 		proj=Proj.Transformation(MA_preset=2,lon0=lon0)
 	
 		fi0=Figure(size=(900,600),fontsize=24)
-		ax0=Axis(fi0[1,1],backgroundcolor = :gray20,
-			title="Distribution of Ocean Observing Platforms")
+		ax0=Axis(fi0[1,1],title="Distribution of Ocean Observing Platforms")
 		pr_ax=MeshArrays.ProjAxis(ax0; proj=proj,lon0=lon0)
 		lines!(pr_ax,polygons=pol;color=:white, linewidth = 0.5)
 	
 		sc1=scatter!(pr_ax,argo_operational.lon,argo_operational.lat,
-			markersize=6.0,label="Argo (operational)",color=:deepskyblue)
+			markersize=6.0,label="Argo (operational)" => (; markersize = 15),color=:deepskyblue)
 		sc2=scatter!(pr_ax,argo_planned.lon,argo_planned.lat,
-			markersize=6.0,label="Argo (planned)",color=:violet)
+			markersize=6.0,label="Argo (planned)" => (; markersize = 15),color=:violet)
 		sc3=scatter!(pr_ax,drifter_operational.lon,drifter_operational.lat,
-			markersize=8.0,label="Drifter",color=:green1)
+			markersize=8.0,label="Drifter" => (; markersize = 15),color=:green1)
 		sc4=scatter!(pr_ax,more_operational.lon,more_operational.lat,
-			markersize=12.0,label=more_platform_name,color=:red,marker=:star5)
+			markersize=12.0,label=more_platform_name => (; markersize = 15),color=:red,marker=:star5)
 		
 		MeshArrays.grid_lines!(pr_ax;color=:yellow,linewidth=0.5)
 
-		Legend(fi0[2, 1],[sc1,sc2,sc3,sc4],[sc1.label,sc2.label,sc3.label,sc4.label],
-			orientation = :horizontal, fontsize=16)
-		
+#		Legend(fi0[2, 1],[sc1,sc2,sc3,sc4],[sc1.label,sc2.label,sc3.label,sc4.label],
+#			orientation = :horizontal, fontsize=16)
+        Legend(fi0[2, 1],ax0,orientation = :horizontal, fontsize=16)
+
 		fi0
 	end
 
-	with_theme(demofigure, theme_dark())
+	fig1=with_theme(demofigure, theme_black())
 end
 
 # ╔═╡ 00000000-0000-0000-0000-000000000001
@@ -168,7 +168,6 @@ CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 DataDeps = "124859b0-ceae-595e-8997-d05f6a7a8dfe"
 GeoJSON = "61d90e0f-e114-555e-ac52-39dfb47a3ef9"
 MeshArrays = "cb8c808f-1acf-59a3-9d2b-6e38d009f683"
-OceanRobots = "0b51df41-3294-4961-8d23-db645e32016d"
 PlutoUI = "7f904dfe-b85e-4ff6-b463-dae2292396a8"
[…]
