+++ paper/paper.tex
@@ -63,7 +63,7 @@ A fraction of the geospatial data obtained by surface ocean drifters in the past
 
 \begin{table}
 \tbl{Ocean observing systems targeted by \pkg{OceanRobots.jl}, and associated data structures (or blank if not yet implemented).}{
-\begin{tabular}{|l||r|}\hline
+\begin{tabular}{||l||r||}\hline
 Platform Type & Data Structure\\\hline\hline
 surface drifters & \struct{SurfaceDrifter}, \struct{CloudDrift} \\
 drifting profilers & \struct{ArgoFloat} \\
@@ -71,7 +71,7 @@ moored buoys & \struct{OceanSite}, \struct{NOAAbuoy} \\
 sea gliders & \struct{Gliders} \\
 expendable bathy-thermographs & \tbd \\
 sail drones & \tbd \\
-\hline
+\hline\hline
 research vessel data & \struct{CCHDO} \\
 ships of opportunity data & \tbd \\
 marine mammals data & \tbd \\\hline
