### 2024-11-29 #164 @gaelforget
> Checking the PDF I see some minor issues:
>
> * code 2 is split which is sub-optimal
> * section headers 3. and 6. are right before end of page
> * Fig. 1 legend (coloured dots) is really hard to interpret (colour dot could be lager)
> * Style of Table (if could be improved)
> * Coloured boxes around hyperlinks (if could be removed)
>
> If possible to address those styling issues, that would be excellent

Thanks for suggesting these. Should be all set now.
