+++ README.md
@@ -1,10 +1,10 @@
 # OceanRobots.jl
 
-[![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://JuliaOcean.github.io/OceanRobots.jl/stable)
-[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://JuliaOcean.github.io/OceanRobots.jl/dev)
 [![CI](https://github.com/JuliaOcean/OceanRobots.jl/actions/workflows/ci.yml/badge.svg)](https://github.com/JuliaOcean/OceanRobots.jl/actions/workflows/ci.yml)
 [![Codecov](https://codecov.io/gh/JuliaOcean/OceanRobots.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/JuliaOcean/OceanRobots.jl)
+[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://JuliaOcean.github.io/OceanRobots.jl/dev)
 
+[![DOI](https://proceedings.juliacon.org/papers/10.21105/jcon.00164/status.svg)](https://doi.org/10.21105/jcon.00164)
 [![DOI](https://zenodo.org/badge/352859934.svg)](https://zenodo.org/badge/latestdoi/352859934)
  
 This package can be used to access, analyze, process, and simulate data generated by _ocean robots_. These ocean observing platforms collect observations in the field, and allow us to monitor climate.
